% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glotto_anderberg_dissimilarity.R
\name{anderberg_dissimilarity}
\alias{anderberg_dissimilarity}
\title{Calculate Anderberg dissimilarity between two objects}
\usage{
anderberg_dissimilarity(glottodata, i, j, type, weights)
}
\arguments{
\item{glottodata}{A dataframe}

\item{i}{An integer}

\item{j}{An integer}
}
\value{
A numeric value
}
\description{
Calculate Anderberg dissimilarity between two objects
}
