% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_idunique}
\alias{check_idunique}
\title{Check wether all IDs are unique}
\usage{
check_idunique(data, idcol, messages = TRUE)
}
\arguments{
\item{data}{User glottodata}

\item{idcol}{Column index or name with IDs (glottocodes, glottosubcodes, or glottosubsubcodes)}

\item{messages}{TRUE/FALSE: should messages be generated?}
}
\value{

}
\description{
Check wether all IDs are unique
}
