% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkuserdata.R
\name{checkdata_glottocodes}
\alias{checkdata_glottocodes}
\title{Check whether all IDs are valid glottocodes.}
\usage{
checkdata_glottocodes(data, idcol, messages = TRUE)
}
\arguments{
\item{idcol}{}
}
\value{
Besides diagnostic messages, this function invisibly returns TRUE if check is passed (all IDs are unique) and FALSE otherwise
}
\description{
This is a wrapper around glottocode_exists
}
