% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocheck.R
\name{checkdata_glottocol}
\alias{checkdata_glottocol}
\title{Check whether there is one column named 'glottocode'}
\usage{
checkdata_glottocol(glottodata)
}
\arguments{
\item{glottodata}{A glottodata table}
}
\value{

}
\description{
Check whether there is one column named 'glottocode'
}
\keyword{internal}
