% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocheck.R
\name{checkdata_glottosubcodes}
\alias{checkdata_glottosubcodes}
\title{Check whether all IDs are valid glottosubcodes.}
\usage{
checkdata_glottosubcodes(glottosubdata)
}
\arguments{
\item{glottosubdata}{glottosubdata}
}
\value{

}
\description{
This is a wrapper around glottosubcode_valid
}
\keyword{internal}
