% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocheck.R
\name{checkdata_glottosubcol}
\alias{checkdata_glottosubcol}
\title{Check whether there is one column named 'glottosubcode'}
\usage{
checkdata_glottosubcol(glottosubdata)
}
\arguments{
\item{glottosubdata}{A glottodata list}
}
\value{

}
\description{
Check whether there is one column named 'glottosubcode'
}
\keyword{internal}
