% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocheck.R
\name{checkdata_idmissing}
\alias{checkdata_idmissing}
\title{Check whether rows have missing IDs.}
\usage{
checkdata_idmissing(data, id)
}
\arguments{
\item{data}{User glottodata}

\item{id}{Column index or name with IDs (glottocodes, glottosubcodes, or glottosubsubcodes)}
}
\value{
Besides diagnostic messages, this function invisibly returns TRUE if check is passed (no missing IDs) and FALSE otherwise
}
\description{
Check whether rows have missing IDs.
}
\examples{
suppressMessages(check_idmissing(data = data, id = 1))
}
\keyword{internal}
