% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkdata_idpresent}
\alias{checkdata_idpresent}
\title{Check whether rows have missing IDs.}
\usage{
checkdata_idpresent(data, idcol)
}
\arguments{
\item{data}{User glottodata}

\item{idcol}{Column index or name with IDs (glottocodes, glottosubcodes, or glottosubsubcodes)}
}
\value{
Besides diagnostic messages, this function invisibly returns TRUE if check is passed (no missing IDs) and FALSE otherwise
}
\description{
Check whether rows have missing IDs.
}
\examples{
suppressMessages(check_idmissing(data = data, idcol = 1))
}
