% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkuserdata.R
\name{checkdata_idunique}
\alias{checkdata_idunique}
\title{Check whether all IDs are unique}
\usage{
checkdata_idunique(data, idcol)
}
\arguments{
\item{data}{User glottodata}

\item{idcol}{Column index or name with IDs (glottocodes, glottosubcodes, or glottosubsubcodes)}
}
\value{
Besides diagnostic messages, this function invisibly returns TRUE if check is passed (all IDs are unique) and FALSE otherwise
}
\description{
Some unicorns might be found...
}
