% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkuserdata.R
\name{checkdata_lscolcount}
\alias{checkdata_lscolcount}
\title{Check whether number of columns are identical across all glots (a language data.frame) in a list}
\usage{
checkdata_lscolcount(langlist)
}
\arguments{
\item{langlist}{}
}
\value{

}
\description{
Check whether number of columns are identical across all glots (a language data.frame) in a list
}
