% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkdata_twolevels}
\alias{checkdata_twolevels}
\title{Check whether each variable has at least two levels (excluding NA).}
\usage{
checkdata_twolevels(data)
}
\arguments{
\item{data}{}
}
\value{
Besides diagnostic messages, this function invisibly returns TRUE if check is passed (all IDs are unique) and FALSE otherwise
}
\description{
This function checks whether each variable has at least two levels (excluding NA). Use function checkall_varlevels to get an overview of the levels in each variable.
}
\examples{
suppressMessages(checkall_twolevels(data = data))
}
