% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkuserdata.R
\name{checkglottodata}
\alias{checkglottodata}
\title{Quality check of user-provided glottodata}
\usage{
checkglottodata(data = NULL, idcol = NULL, show = TRUE)
}
\arguments{
\item{data}{User-provided glottodata}

\item{idcol}{Column index or column name with IDs (glottocodes, glottosubcodes, or glottosubsubcodes)}

\item{show}{If TRUE (default) a data viewer will be opened to showing the levels of each variable (including NAs)}
}
\value{

}
\description{
Go through a user-provided glottodataset and check:
- No missing IDs
- No duplicate IDs
- All variables have at least two levels
- All glottocodes are valid
}
\examples{
dpath <- "C:/Users/sjnor/surfdrive/Shared/SAPPHIRE/Output/Presentations/Isolates_WS-Athens/data/dbase_isolates_V2.xlsx"
glottodata <- get_sheetdata(path = dpath, sheets = 1)
checkglottodata(data = glottodata, idcol = 1)

checkglottodata(data = glottodata, idcol = "glottocode")
lapply(glottodatalist, checkglottodata, idcol = "glottosubcode")
}
