% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkglottodata.R
\name{checkglottosubdata}
\alias{checkglottosubdata}
\title{Quality check of user-provided glottosubdata}
\usage{
checkglottosubdata(glottosubdata, show = TRUE)
}
\arguments{
\item{glottosubdata}{}
}
\value{

}
\description{
Quality check of user-provided glottosubdata
}
\examples{
glottosubdata <- get_glottodata(meta = FALSE, dummy = "glottosubdata")
glottosubdata <- join_glottodata(glottosubdata)
checkglottosubdata(glottosubdata)

Better to join_glottodata first, instead of the following approach (because checks only within each language for duplicates etc.)
lapply(glottosubdata, checkglottosubdata)
}
