% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanglottodata.R
\name{cleandata_recodemissing}
\alias{cleandata_recodemissing}
\title{Reclass missing values to NA}
\usage{
cleandata_recodemissing(glottodata, rec = NULL)
}
\arguments{
\item{glottodata}{User-provided glottodata}

\item{rec}{Optional, additional values to recode to NA}
}
\value{

}
\description{
Reclass missing values to NA
}
\examples{
glottodata <- get_glottodata(meta = FALSE)
cleandata_recodemissing(glottodata, rec = "N")
}
