% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanglottodata.R
\name{cleanglottodata}
\alias{cleanglottodata}
\title{Clean glottodata}
\usage{
cleanglottodata(glottodata, structure = NULL)
}
\arguments{
\item{glottodata}{User-provided glottodata (list or data.frame)}
}
\value{

}
\description{
Clean glottodata
}
\examples{
# list:
glottodata_list <- get_glottodata()
glottodata_list <- cleanglottodata(glottodata_list)
glottodata_list[[1]] <- cleanglottodata(glottodata_list[[1]])

# data.frame:
glottodata <- get_glottodata(meta = FALSE)
glottodata <- cleanglottodata(glottodata)
}
