% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{contransform_lonlat}
\alias{contransform_lonlat}
\title{Conditionally transforms coordinates to lonlat}
\usage{
contransform_lonlat(data)
}
\arguments{
\item{data}{Spatial object with any crs}
}
\value{
Spatial object transformed to lonlat
}
\description{
If CRS of a spatial object is not longitude latitude it will be converted.
}
