% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodist.R
\name{countwithinradius}
\alias{countwithinradius}
\title{Count number of points within a radius}
\usage{
countwithinradius(dist, r)
}
\arguments{
\item{dist}{dist object}

\item{r}{search radius (in same units as \code{dist} object)}
}
\value{
Numeric vector with distances
}
\description{
Count number of points within a radius
}
\examples{
countwithinradius(dist)
}
\seealso{
Other <geodist><glottodist>: 
\code{\link{meandist}()},
\code{\link{nearestdistance}()},
\code{\link{nearestfeature}()}
}
\concept{<geodist><glottodist>}
