% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createglottodata.R
\name{create_emptyglottodata}
\alias{create_emptyglottodata}
\title{Create empty glottodataset for specified glottocodes and features.}
\usage{
create_emptyglottodata(glottocodes, variables, path)
}
\arguments{
\item{path}{}

\item{features}{Either a vector with variable names, or a number indicating the number of variable names to be generated.}
}
\value{

}
\description{
Create empty glottodataset for specified glottocodes and features.
}
