% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createglottodata.R
\name{create_glottosubcodes}
\alias{create_glottosubcodes}
\title{Create glottosubcodes}
\usage{
create_glottosubcodes(glottocode, groups = NULL, n)
}
\arguments{
\item{glottocode}{A glottocode}

\item{groups}{Vector of group names (optional)}

\item{n}{Number of records in each group}
}
\value{

}
\description{
Generate glottosubcodes for a single glottocode
}
\examples{
create_glottosubcodes(glottocode = "yucu1253", groups = c("a", "b"), n = 5)
}
