% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocreate.R
\name{create_structuretable}
\alias{create_structuretable}
\title{create structure table for glottodata}
\usage{
create_structuretable(glottocodes, varnames = NULL)
}
\arguments{
\item{glottocodes}{Character vector of glottocodes}

\item{varnames}{Character vector of variable names}
}
\value{

}
\description{
create structure table for glottodata
}
\examples{
create_structuretable(glottocodes = c("yucu1253", "tani1257"), varnames = c("var001", "var002", "var003"))
}
