% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotools.R
\name{extractraster}
\alias{extractraster}
\title{Extract raster values from glottopoints or glottopols}
\usage{
extractraster(glottodata, geodata, radius = NULL, fun = NULL, add = TRUE)
}
\arguments{
\item{glottodata}{Spatial vector with glottopoints or glottopols}

\item{geodata}{Raster}

\item{radius}{Radius (km) from which raster values should be extracted.}

\item{fun}{One of mean, median, min, max, sd, modal. Alternatively, a user-defined function can be specified, as in fun = function(x){mean(x,na.rm=T)}}

\item{add}{By default, extracted values are added to glottodata.}
}
\value{

}
\description{
Raster values can be extracted at glottopoint locations, from a search radius surrounding glottopoints, or from glottopols
}
\seealso{
Other <geodata><geotools>: 
\code{\link{extractgeodata}()},
\code{\link{mergevec}()},
\code{\link{rasterstack}()}
}
\concept{<geodata><geotools>}
\keyword{internal}
