% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodist.R
\name{geodist}
\alias{geodist}
\title{Pairwise distances between spatial objects}
\usage{
geodist(points, lines = NULL, fun = NULL, label = NULL)
}
\arguments{
\item{points}{Spatial points between which distances should be calculated
(sf)}

\item{lines}{Optional, spatial lines to which distances should be calculated}

\item{fun}{function indicating how distances should be calculated. Default is
"bird" (Haversine distance/great circle distance). Other options are:
"leastcost" (least cost distance), "along" (along lines such as rivers or roads), and
"resistance" (commute distance/random walk/drunkard)}

\item{label}{column name that should be used to label points.}
}
\value{

}
\description{
This function calculates pairwise distances between many points, and distances between many points and many lines.
Different distance metrics are possible: least cost distance, river distance,
}
\examples{
glottodata <- glottoget()
glottodataspace <- glottospace(glottodata)
geodist(points = glottodataspace)
}
\seealso{
Other <geodist>: 
\code{\link{nearestline_bird}()},
\code{\link{pointdist_bird}()},
\code{\link{pointdist_topo}()}
}
\concept{<geodist>}
