% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoget.R
\name{geoget}
\alias{geoget}
\title{Get geographic data}
\usage{
geoget(geodata, country = NULL)
}
\arguments{
\item{geodata}{Either a filepath to locally stored geodata, or the name of
geodata to download ("elevation" or "climate"). This calls raster::getData.}
}
\value{

}
\description{
Load spatial data either from local path (supports both raster and vector
formats), or download from remote server.
}
\examples{
From local path (not run):
geoget(geodata = "D:/data/vector.shp")
geoget(geodata = "D:/data/raster.tif")

Download:
geoget(geodata = "elevation", country = "Netherlands")
geoget(geodata = "climate")
}
\seealso{
Other <geodata>: 
\code{\link{geoget_path}()},
\code{\link{geoget_remote}()}
}
\concept{<geodata>}
