% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoget.R
\name{geoget_basemap}
\alias{geoget_basemap}
\title{Get basemap polygon for the entire world or specific region}
\usage{
geoget_basemap(
  crs = NULL,
  attributes = FALSE,
  continent = NULL,
  country = NULL,
  region = NULL,
  union = FALSE,
  plot = FALSE,
  antarctica = FALSE
)
}
\arguments{
\item{crs}{Coordinate Reference System to transform to (default is Eckert IV)}

\item{attributes}{Whether to keep polygon attributes, or return only geometries}

\item{continent}{Optional filter by continent}

\item{country}{Optional filter by country}

\item{region}{Optional filter by region}

\item{union}{Should the polygons be merged into a single polygon with resolved boundaries?}

\item{plot}{Should the resulting polygon(s) be plotted}

\item{antarctica}{Should Antarctica be kept?}
}
\value{

}
\description{
Download, prepare and (optionally) filter Natural Earth data.
}
\examples{
geoget_basemap(plot = TRUE, union = TRUE)
}
\seealso{
Other <geodata>: 
\code{\link{geoget_path}()},
\code{\link{geoget_remote}()},
\code{\link{geoget}()}
}
\concept{<geodata>}
