% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoget.R
\name{geoget_path}
\alias{geoget_path}
\title{Get spatial data from local path}
\usage{
geoget_path(filepath)
}
\arguments{
\item{filepath}{Path to raster (RasterLayer or RasterStack) or vector data}
}
\value{

}
\description{
Open spatial data from a local path (supports both raster and vector formats).
}
\seealso{
Other <geodata>: 
\code{\link{geoget_remote}()},
\code{\link{geoget}()}
}
\concept{<geodata>}
\keyword{internal}
