% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoget.R
\name{geoget_remote}
\alias{geoget_remote}
\title{Download geodata}
\usage{
geoget_remote(download, country = NULL)
}
\arguments{
\item{download}{name of data to download ("climate" or "elevation")}

\item{country}{Required in case of "elevation" to crop the data}
}
\value{

}
\description{
Download geodata
}
\seealso{
Other <geodata>: 
\code{\link{geoget_path}()},
\code{\link{geoget}()}
}
\concept{<geodata>}
\keyword{internal}
