% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottomap.R
\name{geomap}
\alias{geomap}
\title{Map environmental data}
\usage{
geomap(geodata)
}
\arguments{
\item{geodata}{}
}
\value{

}
\description{
Map environmental data
}
\examples{
nl <- geoget(download = "elevation", country = c("Netherlands"))
plotgeodata(nl)
}
\concept{<geodata><glottomap><geotools>}
\keyword{internal}
