% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodata.R
\name{get_geodata}
\alias{get_geodata}
\title{Get spatial data}
\usage{
get_geodata(path)
}
\arguments{
\item{path}{Path to raster or vector data}
}
\value{

}
\description{
Open spatial data from a local path (supports both raster and vector formats).
}
