% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geodata.R
\name{get_geodata_download}
\alias{get_geodata_download}
\title{Download geodata}
\usage{
get_geodata_download(download, country = NULL)
}
\arguments{
\item{download}{name of data to download ("climate" or "elevation")}

\item{country}{In some cases, this is required to crop the data}
}
\value{

}
\description{
Download geodata
}
\seealso{
Other <geodata>: 
\code{\link{extractgeodata}()},
\code{\link{extractraster}()},
\code{\link{get_geodata_path}()},
\code{\link{get_geodata}()},
\code{\link{glottomap_geodata}()}
}
\concept{<geodata>}
\keyword{internal}
