% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_glottobase.R
\name{get_glottobase}
\alias{get_glottobase}
\title{Get glottobase reference data}
\usage{
get_glottobase()
}
\value{

}
\description{
Downloads most recent glottolog data and transforms it. This 'glottobase' is used as reference dataset in several functions.
}
\examples{
glottobase <- get_glottobase()
}
