% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget.R
\name{get_glottodata}
\alias{get_glottodata}
\title{Load glottodata from an excel file}
\usage{
get_glottodata(
  filename = NULL,
  meta = FALSE,
  simplify = TRUE,
  create = "glottodata"
)
}
\arguments{
\item{filename}{Path to glottodata file with extension (.xlsx .xls .gpkg .shp). If no filename is specified, an artificial dummy dataset will be created.}

\item{meta}{By default, meta sheets are not loaded. Use meta=TRUE if you want to include them.}

\item{simplify}{By default, if only one sheet is loaded, the data will be returned as a data.frame (instead of placing the data inside a list of length 1)}

\item{create}{In case 'filename' is not specified, artificial dummy data will be created in glottodata format (specify create = "glottosubdata" to create data in glottosubdata format)}
}
\value{

}
\description{
Load glottodadata/glottosubdata from a file, or create artificial dummy data.
}
\examples{
get_glottodata()
get_glottodata(filename = "glottodata.xlsx")
get_glottodata(filename = "glottodata.gpkg")
}
\seealso{
glottosave

Other <glottodata>: 
\code{\link{glottoget}()},
\code{\link{glottosave}()}
}
\concept{<glottodata>}
