% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_glottobase.R
\name{get_glottospace}
\alias{get_glottospace}
\title{Get glottospace reference data}
\usage{
get_glottospace()
}
\value{

}
\description{
Get most recent glottolog data and turn it into the most elemental geoglot object (i.e. glottocodes + geometry column). This 'glottospace' is used as reference dataset in several functions.
}
\examples{
glottospace <- get_glottospace()
}
\seealso{
glottodata_addcoords
}
