% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodata.R
\name{get_rasveclist}
\alias{get_rasveclist}
\title{Try to open raster/vector data and return list of two}
\usage{
get_rasveclist(path)
}
\arguments{
\item{path}{Path to raster or vector data}
}
\value{
A list of two
}
\description{
Try to open raster/vector data and return list of two
}
