% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottodata.R
\name{get_sheetdata}
\alias{get_sheetdata}
\title{Import sheets from an excel file (.xls or .xlsx). This is a wrapper around excel_sheets.
Use the function \code{"get_sheetnames"} to show all sheetnames in an excel file (without importing them).}
\usage{
get_sheetdata(filepath, sheets = NULL)
}
\arguments{
\item{filepath}{Path to excel file (.xls or .xlsx)}

\item{sheets}{Character vector of sheetnames to import. Leave empty in case you want to import all sheets.}
}
\value{
List with tibble per sheet
}
\description{
Import sheets from an excel file (.xls or .xlsx). This is a wrapper around excel_sheets.
Use the function \code{"get_sheetnames"} to show all sheetnames in an excel file (without importing them).
}
\examples{
get_sheetdata(filepath = "glottodata.xlsx")
get_sheetdata(filepath = "glottodata.xlsx", sheets = "abcd1234")
}
