% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottodata.R
\name{get_sheetnames}
\alias{get_sheetnames}
\title{List names of all sheets in an excel file (.xls or .xlsx). This is a wrapper around excel_sheets.
Use the function \code{"get_sheetdata"}  to import data.}
\usage{
get_sheetnames(filepath)
}
\arguments{
\item{path}{Path to excel file}
}
\value{
Vector of names
}
\description{
List names of all sheets in an excel file (.xls or .xlsx). This is a wrapper around excel_sheets.
Use the function \code{"get_sheetdata"}  to import data.
}
\examples{
get_sheetnames(path = "glottodata.xlsx")
get_sheetnames(path = "glottodata.xls")
}
