% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialdata.R
\name{get_spatialdata}
\alias{get_spatialdata}
\title{Get spatial data}
\usage{
get_spatialdata(path)
}
\arguments{
\item{path}{Path to raster or vector data}
}
\value{

}
\description{
Open spatial data from a local path (supports both raster and vector formats).
}
