% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcldf.R
\name{get_spec}
\alias{get_spec}
\title{Returns a dataset specification for `readr` or `vroom` from the metadata file}
\usage{
get_spec(dt)
}
\arguments{
\item{dt}{a column schema spec}
}
\value{
a `readr` column spec
}
\description{
Returns a dataset specification for `readr` or `vroom` from the metadata file
}
