% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcldf.R
\name{get_table_schema}
\alias{get_table_schema}
\title{Extracts the table schema from the metadata schema}
\usage{
get_table_schema(schema)
}
\arguments{
\item{schema}{the metadata schema.}
}
\value{
A column schema
}
\description{
Extracts the table schema from the metadata schema
}
