% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcldf.R
\name{get_tablename}
\alias{get_tablename}
\title{Convert a CLDF URL tablename to a short tablename}
\usage{
get_tablename(tbl)
}
\arguments{
\item{tbl}{the tablename.}
}
\value{
A string
}
\description{
Convert a CLDF URL tablename to a short tablename
}
\examples{
get_tablename("languages.csv")
}
