% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottosearch.R
\name{glot_exists_one}
\alias{glot_exists_one}
\title{Check whether one glot exists in glottolog.}
\usage{
glot_exists_one(find, columns)
}
\value{
Logical: TRUE/FALSE
}
\description{
Check whether one glot exists in glottolog.
}
