% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{glotto_search}
\alias{glotto_search}
\title{Title}
\usage{
glotto_search(glottodata = "glottolog", find = NULL, partialmatch = TRUE)
}
\arguments{
\item{name}{Name of a language, family, or dialect}

\item{partial}{Whether partial matches should be returned besides exact matches.}

\item{isocode}{Three letter isocode(s): "abc" or c("abc", "abd")}

\item{glottocode}{Glottocode(s): "abcd1234" or c("abcd1234", "abcd1235")}

\item{webpage}{Logical, whether}

\item{logical}{}
}
\value{

}
\description{
Title
}
