% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glottobase}
\alias{glottobase}
\title{Load default glottolog reference data.
Loads glottolog data which is used in other functions as a reference. For example, to check wether a glottocode exists.
Glottobase: glottolog dataset used as a reference/lookup}
\format{
A data frame with XX rows and XX variables:
\describe{
  \item{var1}{explain}
  \item{var2}{explain}
  ...
}
}
\source{
\url{http://www.glottolog.org/
}
}
\usage{
glottobase
}
\description{
Glottolog dataset.
}
\keyword{datasets}
