% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottobooster.R
\name{glottobooster_addisolates}
\alias{glottobooster_addisolates}
\title{Glottolog: add isolates}
\usage{
glottobooster_addisolates(glottologdata)
}
\arguments{
\item{glottologdata}{data from \href{https://glottolog.org/}{glottolog}, can be downloaded with \code{\link{glottoget("glottolog")}}}
}
\value{

}
\description{
Glottolog: add isolates
}
\seealso{
Other <glottobooster>: 
\code{\link{glottobooster_L1only}()},
\code{\link{glottobooster_addfamilyname}()},
\code{\link{glottobooster_addfamilysizerank}()},
\code{\link{glottobooster_addfamilysize}()},
\code{\link{glottobooster_rmartifam}()},
\code{\link{glottobooster_rmbookkeeping}()},
\code{\link{glottobooster_rmdialects}()},
\code{\link{glottobooster_rmfamilies}()},
\code{\link{glottobooster_rmsignlangs}()},
\code{\link{glottoboosterflex}()},
\code{\link{glottobooster}()}
}
\concept{<glottobooster>}
\keyword{internal}
