% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottobooster.R
\name{glottobooster_L1only}
\alias{glottobooster_L1only}
\title{Glottolog: keep only L1 languages}
\usage{
glottobooster_L1only(glottologdata)
}
\arguments{
\item{glottologdata}{data from \href{https://glottolog.org/}{glottolog}, can be downloaded with glottoget("glottolog")}
}
\description{
Filter out only L1 languages and drop 'category' column.
}
\seealso{
Other <glottobooster>: 
\code{\link{glottobooster_addfamilyname}()},
\code{\link{glottobooster_addfamilysizerank}()},
\code{\link{glottobooster_addfamilysize}()},
\code{\link{glottobooster_addisolates}()},
\code{\link{glottobooster_rmartifam}()},
\code{\link{glottobooster_rmbookkeeping}()},
\code{\link{glottobooster_rmdialects}()},
\code{\link{glottobooster_rmfamilies}()},
\code{\link{glottobooster_rmsignlangs}()},
\code{\link{glottoboosterflex}()},
\code{\link{glottobooster}()}
}
\concept{<glottobooster>}
\keyword{internal}
