% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottobooster.R
\name{glottobooster_rmsignlangs}
\alias{glottobooster_rmsignlangs}
\title{Glottolog: remove sign languages}
\usage{
glottobooster_rmsignlangs(glottologdata)
}
\arguments{
\item{glottologdata}{data from \href{https://glottolog.org/}{glottolog}, can be downloaded with glottoget("glottolog")}
}
\description{
Glottolog: remove sign languages
}
\seealso{
Other <glottobooster>: 
\code{\link{glottobooster_L1only}()},
\code{\link{glottobooster_addfamilyname}()},
\code{\link{glottobooster_addfamilysizerank}()},
\code{\link{glottobooster_addfamilysize}()},
\code{\link{glottobooster_addisolates}()},
\code{\link{glottobooster_rmartifam}()},
\code{\link{glottobooster_rmbookkeeping}()},
\code{\link{glottobooster_rmdialects}()},
\code{\link{glottobooster_rmfamilies}()},
\code{\link{glottoboosterflex}()},
\code{\link{glottobooster}()}
}
\concept{<glottobooster>}
\keyword{internal}
