% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocheck.R
\name{glottocheck}
\alias{glottocheck}
\title{Quality check of user-provided glottodata or glottosubdata}
\usage{
glottocheck(glottodata, diagnostic = TRUE)
}
\arguments{
\item{glottodata}{User-provided glottodata}

\item{diagnostic}{If TRUE (default) a data viewer will be opened to show the levels of each variable (including NAs), and a data coverage plot will be shown.}
}
\value{

}
\description{
This function first checks whether a dataset is glottodata or glottosubdata, and depending on the result calls checkglottodata or checkglottosubdata.
}
\details{
It subsequently checks whether:
- one column exists with the name "glottocode"
- there are rows without a glottocode  (missing IDs)
- there are rows with duplicated glottocodes (duplicate IDs)
- all variables have at least two levels
- all glottocodes are valid
}
\examples{
glottocheck(glottodata)
}
\seealso{
Other <glottocheck>: 
\code{\link{checkglottodata}()},
\code{\link{checkglottosubdata}()},
\code{\link{glottocheck_hasmeta}()}
}
\concept{<glottocheck>}
