% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocheck.R
\name{glottocheck_colmissing}
\alias{glottocheck_colmissing}
\title{Check whether columns have missing data}
\usage{
glottocheck_colmissing(data, id, diagnostic = FALSE, rm.na = TRUE)
}
\arguments{
\item{data}{A dataset}

\item{id}{Column name or index with unique id's}

\item{diagnostic}{Whether diagnostic messages should be shown}

\item{rm.na}{Whether rows without id should be removed.}
}
\value{

}
\description{
Provides a message in case there are missing data, otherwise returns NULL
}
\keyword{internal}
