% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocheck.R
\name{glottocheck_data}
\alias{glottocheck_data}
\title{Quality check of user-provided glottodata}
\usage{
glottocheck_data(glottodata, diagnostic = TRUE)
}
\arguments{
\item{glottodata}{User-provided glottodata}

\item{diagnostic}{If TRUE (default) a data viewer will be opened to show the levels of each variable (including NAs), and a data coverage plot will be shown.}
}
\description{
Go through a user-provided glottodataset and check whether:
\itemize{
  \item one column exists with the name "glottocode"
  \item there are rows without a glottocode  (missing IDs)
  \item there are rows with duplicated glottocodes (duplicate IDs)
  \item all variables have at least two levels
  \item all glottocodes are valid
}
}
\examples{
glottodata <- glottoget("demodata")
glottocheck_data(glottodata, diagnostic = FALSE)
}
\seealso{
Other <glottocheck>: 
\code{\link{glottocheck_hasmeta}()},
\code{\link{glottocheck_subdata}()},
\code{\link{glottocheck}()}
}
\concept{<glottocheck>}
\keyword{internal}
