% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocheck.R
\name{glottocheck_glottocodes}
\alias{glottocheck_glottocodes}
\title{Check whether all IDs are valid glottocodes.}
\usage{
glottocheck_glottocodes(glottodata)
}
\arguments{
\item{glottodata}{User-provided glottodata}
}
\value{
Besides diagnostic messages, this function invisibly returns TRUE if check is passed (all IDs are unique) and FALSE otherwise
}
\description{
This is a wrapper around glottocode_exists
}
\keyword{internal}
