% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocheck.R
\name{glottocheck_glottosubcodes}
\alias{glottocheck_glottosubcodes}
\title{Check whether all IDs are valid glottosubcodes.}
\usage{
glottocheck_glottosubcodes(glottosubdata)
}
\arguments{
\item{glottosubdata}{glottosubdata}
}
\description{
This is a wrapper around glottosubcode_valid
}
\keyword{internal}
