% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocheck.R
\name{glottocheck_hasmeta}
\alias{glottocheck_hasmeta}
\alias{glottocheck_hasstructure}
\title{Check whether glottodata contains metadata}
\usage{
glottocheck_hasmeta(glottodata)
}
\arguments{
\item{glottodata}{glottodata}
}
\description{
In fact, this function only checks whether glottodata contains a structure table, because the structure table is the only table that is required by some glottospace functions. All other tables are for humans, not computers;-).
}
\examples{
glottocheck_hasmeta(glottodata)
}
\seealso{
Other <glottocheck>: 
\code{\link{glottocheck_data}()},
\code{\link{glottocheck_subdata}()},
\code{\link{glottocheck}()}
}
\concept{<glottocheck>}
\keyword{internal}
