% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocheck.R
\name{glottocheck_metadata}
\alias{glottocheck_metadata}
\title{Check metadata of glottodata}
\usage{
glottocheck_metadata(glottodata)
}
\arguments{
\item{glottodata}{glottodata}
}
\description{
Check metadata of glottodata
}
\examples{
glottodata <- glottoget("demodata", meta = TRUE)
glottocheck_metadata(glottodata)
}
\keyword{internal}
