% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocheck.R
\name{glottocheck_subdata}
\alias{glottocheck_subdata}
\title{Quality check of user-provided glottosubdata}
\usage{
glottocheck_subdata(glottosubdata, diagnostic = TRUE)
}
\arguments{
\item{glottosubdata}{User-provided glottosubdata}

\item{diagnostic}{If TRUE (default) a data viewer will be opened to show the levels of each variable (including NAs), and a data coverage plot will be shown.}
}
\description{
Quality check of user-provided glottosubdata
}
\examples{
glottosubdata <- glottocreate_demosubdata()
glottocheck_subdata(glottosubdata)
}
\seealso{
Other <glottocheck>: 
\code{\link{glottocheck_data}()},
\code{\link{glottocheck_hasmeta}()},
\code{\link{glottocheck}()}
}
\concept{<glottocheck>}
\keyword{internal}
