% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocheck.R
\name{glottocheck_twolevels}
\alias{glottocheck_twolevels}
\title{Check whether each variable has at least two levels (excluding NA).}
\usage{
glottocheck_twolevels(data)
}
\arguments{
\item{data}{}
}
\value{
Besides diagnostic messages, this function invisibly returns TRUE if check is passed (all variables have at least two levels) and FALSE otherwise
}
\description{
This function checks whether each variable has at least two levels (excluding NA). Use function glottocheck_varlevels to get an overview of the levels in each variable.
}
\examples{
suppressMessages(glottocheck_twolevels(data = data))
}
\keyword{internal}
