% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocheck.R
\name{glottocheck_varlevels}
\alias{glottocheck_varlevels}
\title{Show levels of all variables.}
\usage{
glottocheck_varlevels(data)
}
\arguments{
\item{data}{}
}
\value{
Opens a data viewer showing the levels of each variable (including NA)
}
\description{
This function shows the levels of all variables and is mainly intended to provide users with a quick overview of their data.
}
\examples{
suppressMessages(glottocheck_varlevels(data = data))
}
\keyword{internal}
