% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoclean.R
\name{glottoclean}
\alias{glottoclean}
\title{Clean glottodata}
\usage{
glottoclean(glottodata, structure = NULL)
}
\arguments{
\item{glottodata}{User-provided glottodata (list or data.frame)}
}
\value{

}
\description{
Clean glottodata
}
\examples{
# list:
glottodata_list <- glottoget_path()
glottodata_list <- glottoclean(glottodata_list)
glottodata_list[[1]] <- glottoclean(glottodata_list[[1]])

# data.frame:
glottodata <- glottoget_path(meta = TRUE)
glottodata <- glottoclean(glottodata)
}
\seealso{
Other <glottoclean>: 
\code{\link{glottoclean_recodemissing}()}
}
\concept{<glottoclean>}
