% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoclean.R
\name{glottoclean_colnamerepair}
\alias{glottoclean_colnamerepair}
\title{Fix colnames of excel files in which colnames refer to another cell}
\usage{
glottoclean_colnamerepair(glottodata)
}
\arguments{
\item{glottodata}{glottodata or glottosubdata}
}
\description{
Fix colnames of excel files in which colnames refer to another cell
}
