% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoclean.R
\name{glottoclean_recodelogical}
\alias{glottoclean_recodelogical}
\title{Recode logical columns to TRUE/FALSE}
\usage{
glottoclean_recodelogical(glottodata, structure)
}
\arguments{
\item{glottodata}{User-provided glottodata}

\item{structure}{A glottodata structure table}
}
\value{

}
\description{
Recode logical columns to TRUE/FALSE
}
\keyword{internal}
