% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoclean.R
\name{glottoclean_selectsample}
\alias{glottoclean_selectsample}
\title{Subset glottodata based on sample table or glottocodes}
\usage{
glottoclean_selectsample(glottodata)
}
\arguments{
\item{glottodata}{glottodata or glottosubdata}
}
\value{
A simplified version of glottodata containing only the selected glottocodes.
}
\description{
Subset glottodata based on sample table or glottocodes
}
\examples{
glottodata <- glottoget("demodata", meta = TRUE)
}
