% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocode.R
\name{glottocode}
\alias{glottocode}
\title{Get detailed info for a specific glottocode}
\usage{
glottocode(glottocode)
}
\arguments{
\item{glottocode}{}
}
\value{

}
\description{
This function does three things for a glottocode:
- It plots its location on the globe
- It opens the associated webpage on www.glottolog.org
- It checks whether it exists in glottolog.
}
\examples{
glottocode("yucu1253")
}
