% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocode.R
\name{glottocode_location}
\alias{glottocode_location}
\title{Show location of glottocode on globe}
\usage{
glottocode_location(glottocode)
}
\arguments{
\item{glottocode}{}
}
\value{

}
\description{
Show location of glottocode on globe
}
\examples{
glottocode_location("yucu1253")
}
\keyword{internal}
