% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottosearch.R
\name{glottocode_online}
\alias{glottocode_online}
\title{Open url in web browser for glottocode}
\usage{
glottocode_online(glottocode)
}
\arguments{
\item{glottocode}{}
}
\value{

}
\description{
Open url in web browser for glottocode
}
\examples{
glottocode_online("yucu1253")
}
