% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocolors.R
\name{glottocolpal}
\alias{glottocolpal}
\title{Color palettes for maps and figures}
\usage{
glottocolpal(palette, ncolr = NULL, rev = FALSE, alpha = NULL, show = FALSE)
}
\arguments{
\item{palette}{Name of color palette, use glottocolpal("all") to see the options}

\item{ncolr}{Number of colors to be generated (optional)}

\item{rev}{Should the order of colors be reversed?}

\item{alpha}{Values between 0 (transparent) and 1 (opaque)}

\item{show}{Do you want to plot the color palette?}
}
\value{
A vector of colors, optionally plotted.
}
\description{
Color palettes for maps and figures
}
\examples{
glottocolpal(palette = "turbo", show = TRUE)
}
\keyword{internal}
