% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocolors.R
\name{glottocolpal_options}
\alias{glottocolpal_options}
\title{Show which color palettes are available}
\usage{
glottocolpal_options()
}
\value{

}
\description{
Show which color palettes are available
}
\keyword{internal}
