% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoconvert.R
\name{glottoconvert}
\alias{glottoconvert}
\title{glottoconvert}
\usage{
glottoconvert(data, glottocode, var = NULL, ref = NULL, page)
}
\arguments{
\item{data}{Dataset to be transformed into glottodata}

\item{glottocode}{column name or column id with glottocodes}

\item{var}{Unique character string that distinguishes columns with variable names.}

\item{ref}{Unique character string that distinguishes columns with references.}

\item{page}{Unique character string that distinguishes columns with page numbers.}
}
\value{

}
\description{
glottoconvert
}
\examples{
# Load data from path:
# testdata <- glottoget("C:/Users/sjnor/surfdrive/Projecten en schrijfsels/Papers in progress/Yucuna-Tanimuca/data/testdata linguistics/2022_01_16_basic_data.xlsx")
}
