% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoconvert.R
\name{glottoconvert_subdata}
\alias{glottoconvert_subdata}
\title{Convert a linguistic dataset into glottosubdata}
\usage{
glottoconvert_subdata(data, glottocodes = NULL)
}
\arguments{
\item{data}{Any dataset that should be converted into glottosubdata. This will generally be an excel file loaded with glottoget() of which the sheetnames are glottocodes.}

\item{glottocodes}{Optional character vector of glottocodes. If no glottocodes are supplied, glottospace will search for them in the sample table.}
}
\description{
Convert a linguistic dataset into glottosubdata
}
\keyword{internal}
