% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoconvert.R
\name{glottoconvert_subtable}
\alias{glottoconvert_subtable}
\title{Transform a table of a linguistic dataset into a glottosubdata table}
\usage{
glottoconvert_subtable(table, glottosubcolumn = NULL, var)
}
\arguments{
\item{table}{In case dataset consists of multiple tables, indicate which table contains the data that should be converted.}

\item{glottosubcolumn}{Column name with glottosubcodes}

\item{var}{Character string that distinguishes those columns which contain variable names.}
}
\description{
Transform a table of a linguistic dataset into a glottosubdata table
}
