% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoconvert.R
\name{glottoconvert_table}
\alias{glottoconvert_table}
\title{Transform a linguistic dataset into glottodata}
\usage{
glottoconvert_table(
  table,
  glottocolumn = NULL,
  var,
  ref = NULL,
  page = NULL,
  remark = NULL,
  contributor = NULL
)
}
\arguments{
\item{table}{In case dataset consists of multiple tables, indicate which table contains the data that should be converted.}

\item{glottocolumn}{column name or column id with glottocodes}

\item{var}{Character string that distinguishes those columns which contain variable names.}

\item{ref}{Character string that distinguishes those columns which contain references.}

\item{page}{Character string that distinguishes those columns which contain page numbers.}

\item{remark}{Character string that distinguishes those columns which contain remarks.}

\item{contributor}{Character string that distinguishes those columns which contain contributors.}

\item{data}{Dataset to be transformed to glottodata}
}
\description{
Transform a linguistic dataset into glottodata
}
\examples{
\dontrun{
data <- glottoget("userdata.xlsx")
glottodata <- glottoconvert_table(data = data, var = "var_", ref = "ref_",
              page = "pag_", remark = "com_", contributor = "Coder")
glottocheck(glottodata)
glottosave(glottodata)
}
}
\keyword{internal}
