% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocreate.R
\name{glottocreate_addstructure}
\alias{glottocreate_addstructure}
\title{Add structure table to glottodata or glottosubdata}
\usage{
glottocreate_addstructure(glottodata)
}
\arguments{
\item{glottodata}{glottodata or glottosubdata}
}
\value{
glottodata/glottosubdata with a structure table
}
\description{
Add structure table to glottodata or glottosubdata
}
\examples{
glottodata <- glottoget("demodata")
glottocreate_addstructure(glottodata)
}
