% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocreate.R
\name{glottocreate_commentstable}
\alias{glottocreate_commentstable}
\title{create comments table for glottodata}
\usage{
glottocreate_commentstable(glottocodes, varnames)
}
\arguments{
\item{glottocodes}{Character vector of glottocodes}

\item{varnames}{Character vector of variable names}
}
\value{

}
\description{
create comments table for glottodata
}
\keyword{internal}
