% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocreate.R
\name{glottocreate_contributorstable}
\alias{glottocreate_contributorstable}
\title{create contributors table for glottodata}
\usage{
glottocreate_contributorstable(glottocodes, varnames)
}
\arguments{
\item{glottocodes}{Character vector of glottocodes}

\item{varnames}{Character vector of variable names}
}
\value{

}
\description{
create contributors table for glottodata
}
\keyword{internal}
