% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocreate.R
\name{glottocreate_data}
\alias{glottocreate_data}
\title{Create empty glottodata for specified glottocodes and variables.}
\usage{
glottocreate_data(
  glottocodes,
  variables,
  filename = NULL,
  meta = TRUE,
  check = FALSE,
  simplify = TRUE,
  levels = NULL,
  maintainer = NULL,
  email = NULL,
  citation = NULL,
  url = NULL
)
}
\arguments{
\item{glottocodes}{Character vector of glottocodes}

\item{variables}{Either a vector with variable names, or a single number indicating the total number of variable columns to be generated}

\item{filename}{Optional name of excel file where to store glottodata}

\item{meta}{By default, meta tables are created. Use meta=FALSE to exclude them.}

\item{check}{Should glottocodes be checked? Default is FALSE because takes much time to run.}

\item{simplify}{By default, if only one table is loaded, the data will be returned as a data.frame (instead of placing the data inside a list of length 1)}

\item{levels}{Optional character vector with levels across all variables}

\item{maintainer}{Name of the person/organization maintaining the data (optional)}

\item{email}{Email address of maintainer/contact person (optional)}

\item{citation}{How to cite the data (optional)}

\item{url}{Optional url linking to a webpage.}
}
\value{

}
\description{
Output is a list with one data.frame for glottodata (and a number of metatables if meta = TRUE). Output can  also be saved as an excel file.
}
\examples{
glottocreate_data(glottocodes = c("yucu1253", "tani1257"),
                   variables = 3, filename = "glottodata.xlsx")
glottocreate_data(glottocodes = c("yucu1253", "tani1257"),
     variables = 3, filename = "glottodata_simple.xlsx", meta = FALSE)
}
\concept{<glottocreate><glottodata>}
\keyword{internal}
