% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocreate.R
\name{glottocreate_descriptiontable}
\alias{glottocreate_descriptiontable}
\title{create descriptiontable for glottodata}
\usage{
glottocreate_descriptiontable(varnames, levelnames = NULL)
}
\arguments{
\item{varnames}{Character vector of variable names}
}
\value{

}
\description{
create descriptiontable for glottodata
}
\keyword{internal}
