% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocreate.R
\name{glottocreate_glottosubcodes}
\alias{glottocreate_glottosubcodes}
\title{Create glottosubcodes}
\usage{
glottocreate_glottosubcodes(glottocode, groups = NULL, n)
}
\arguments{
\item{glottocode}{A glottocode}

\item{groups}{Vector of group names (optional)}

\item{n}{Number of records in each group}
}
\description{
Generate glottosubcodes for a single glottocode
}
\examples{
glottocreate_glottosubcodes(glottocode = "yucu1253",
groups = c("a", "b"), n = 5)
}
