% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocreate.R
\name{glottocreate_glottotable}
\alias{glottocreate_glottotable}
\title{Create glottotable for glottodata}
\usage{
glottocreate_glottotable(glottocodes, varnames)
}
\arguments{
\item{glottocodes}{Character vector of glottocodes}

\item{varnames}{Character vector of variable names}
}
\description{
Create glottotable for glottodata
}
\keyword{internal}
