% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocreate.R
\name{glottocreate_lookuptable}
\alias{glottocreate_lookuptable}
\title{Create lookup table for glottodata}
\usage{
glottocreate_lookuptable()
}
\description{
Create lookup table for glottodata
}
\keyword{internal}
