% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocreate.R
\name{glottocreate_metatable}
\alias{glottocreate_metatable}
\title{create metatable for glottodata}
\usage{
glottocreate_metatable(varnames)
}
\arguments{
\item{varnames}{Character vector of variable names}
}
\value{

}
\description{
create metatable for glottodata
}
\keyword{internal}
