% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocreate.R
\name{glottocreate_metatables}
\alias{glottocreate_metatables}
\title{Create metatables for glottodata or glottosubdata}
\usage{
glottocreate_metatables(
  glottocodes = NULL,
  glottosubcodes = NULL,
  varnames = NULL,
  levels = NULL,
  maintainer = NULL,
  email = NULL,
  citation = NULL,
  url = NULL
)
}
\arguments{
\item{glottocodes}{Character vector of glottocodes}

\item{varnames}{Character vector of variable names}

\item{levels}{Optional character vector with levels across all variables}

\item{maintainer}{Name of the person/organization maintaining the data (optional)}

\item{email}{Email address of maintainer/contact person (optional)}

\item{citation}{How to cite the data (optional)}

\item{url}{Optional url linking to a webpage.}
}
\value{
a list of metatables
}
\description{
Create metatables for glottodata or glottosubdata
}
\keyword{internal}
