% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocreate.R
\name{glottocreate_readmetable}
\alias{glottocreate_readmetable}
\title{Create readme table for glottodata}
\usage{
glottocreate_readmetable(
  maintainer = NULL,
  email = NULL,
  citation = NULL,
  url = NULL
)
}
\arguments{
\item{maintainer}{maintainer of the data}

\item{email}{email address of maintainer}

\item{citation}{how to cite the data}

\item{url}{url with background information on the data}
}
\value{

}
\description{
Create readme table for glottodata
}
\keyword{internal}
