% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocreate.R
\name{glottocreate_reftable}
\alias{glottocreate_reftable}
\title{create reference table for glottodata}
\usage{
glottocreate_reftable(glottocodes, varnames)
}
\arguments{
\item{glottocodes}{Character vector of glottocodes}

\item{varnames}{Character vector of variable names}
}
\description{
create reference table for glottodata
}
\keyword{internal}
