% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocreate.R
\name{glottocreate_remarkstable}
\alias{glottocreate_remarkstable}
\title{create remarks table for glottodata}
\usage{
glottocreate_remarkstable(glottocodes = NULL, varnames = NULL)
}
\arguments{
\item{glottocodes}{Character vector of glottocodes}

\item{varnames}{Character vector of variable names}
}
\description{
create remarks table for glottodata
}
\keyword{internal}
