% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocreate.R
\name{glottocreate_structuretable}
\alias{glottocreate_structuretable}
\title{Create structure table for glottodata}
\usage{
glottocreate_structuretable(glottocodes, varnames = NULL)
}
\arguments{
\item{glottocodes}{Character vector of glottocodes}

\item{varnames}{Character vector of variable names}
}
\value{

}
\description{
This function creates a new structure table, it can be added to glottodata using glottodata_addtable()
}
\examples{
glottocreate_structuretable(glottocodes = c("yucu1253", "tani1257"), varnames = c("var001", "var002", "var003"))
}
