% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottodata.R
\name{glottodata}
\alias{glottodata}
\title{Download glottolog data and convert it into a glot or geoglot object}
\usage{
glottodata(data = "glottodata_spatial")
}
\arguments{
\item{data}{One of \code{"glottodata_raw"}, \code{"glottodata_tidy"}, or
\code{"glottodata_spatial"}. By default, the downloaded glottolog data is
converted into a geoglot object (sf).}
}
\value{
Either a glot or geoglot object.
}
\description{
Download glottolog data and convert it into a glot or geoglot object
}
\examples{
glottodata <- glottodata()
}
