% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geobooster.R
\name{glottodata_addcoords}
\alias{glottodata_addcoords}
\title{Add coordinates to glottodata}
\usage{
glottodata_addcoords(glottodata)
}
\arguments{
\item{glottodata}{}
}
\value{

}
\description{
Add coordinates to glottodata
}
