% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geobooster.R
\name{glottodata_addcountries}
\alias{glottodata_addcountries}
\alias{glottodata_addcontinents}
\alias{glottodata_addregions}
\title{Add names of countries, continents, and regions to glottodata}
\usage{
glottodata_addcountries(glottodata)
}
\value{

}
\description{
Add names of countries, continents, and regions to glottodata
}
