% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocreate.R
\name{glottodata_addtable}
\alias{glottodata_addtable}
\title{Add a table to glottodata}
\usage{
glottodata_addtable(glottodata, table, name)
}
\arguments{
\item{glottodata}{A glottodata table, or a list of glottodata tables}

\item{table}{A table to be added}

\item{name}{A name for the table}
}
\value{

}
\description{
Add a table to glottodata
}
