% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glotto_anderberg_dissimilarity.R
\name{glottodata_cleaned}
\alias{glottodata_cleaned}
\title{Create a cleaned glottodata fitting for calculating dissimilarities}
\usage{
glottodata_cleaned(glottodata)
}
\arguments{
\item{glottodata}{glottodata or glottosubdata, either with or without structure table.}
}
\value{
A dataframe
}
\description{
Create a cleaned glottodata fitting for calculating dissimilarities
}
\examples{
glottodata <- glottoget("demodata", meta = TRUE)
glottodata_cleaned(glottodata)

glottosubdata <- glottoget("demosubdata", meta = TRUE)
glottodata_cleaned(glottosubdata)
}
