% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottospace.R
\name{glottodata_makespatial}
\alias{glottodata_makespatial}
\alias{glottodata_addcoords}
\title{Make glottodata spatial}
\usage{
glottodata_makespatial(glottodata)
}
\arguments{
\item{glottodata}{User-provided glottodata}
}
\value{
glottodata with a GEOMETRY column
}
\description{
Adds coordinates to glottodata.
If glottodata does not contain a geometry column, it will be added.
}
\examples{
glottodata <- glottoget()
glottodata_makespatial(glottodata)
}
