% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottodist.R
\name{glottodist}
\alias{glottodist}
\title{Calculate distances between languages}
\usage{
glottodist(glottodata, structure = NULL)
}
\arguments{
\item{glottodata}{glottodata or glottosubdata, either with or without structure table.}

\item{structure}{If glottodata is a table without a structure table, you can add it separately. To create a structure table, you should run glottocreate_structuretable() and you can add it with glottocreate_addtable()}
}
\value{
object of class \code{dist}
}
\description{
Calculate distances between languages
}
\examples{
glottodata <- glottoget("demodata", meta = TRUE)
glottodist <- glottodist(glottodata = glottodata)

glottosubdata <- glottoget("demosubdata", meta = TRUE)
glottodist <- glottodist(glottodata = glottosubdata)

}
