% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glotto_anderberg_dissimilarity.R
\name{glottodist_anderberg}
\alias{glottodist_anderberg}
\title{Calculate the Anderberg dissimilarity distance matrix of languages}
\usage{
glottodist_anderberg(glottodata)
}
\arguments{
\item{glottodata}{glottodata or glottosubdata, either with or without structure table.}
}
\value{
object of class \code{dist}
}
\description{
Calculate the Anderberg dissimilarity distance matrix of languages
}
\examples{
glottodata <- glottoget("demodata", meta = TRUE)
glottodist_anderberg(glottodata)

glottosubdata <- glottoget("demosubdata", meta = TRUE)
glottodist_anderberg(glottosubdata)
}
