% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottodist.R
\name{glottodist_subdata}
\alias{glottodist_subdata}
\title{Title}
\usage{
glottodist_subdata(
  glottosubdata,
  metric = NULL,
  index_type = NULL,
  avg_idx = NULL,
  fixed_idx = NULL
)
}
\arguments{
\item{glottosubdata}{an glottosubdata object}

\item{metric}{either "gower" or "anderberg"}

\item{index_type}{either "mc" or "relative" or "fmi"}

\item{avg_idx}{the feature indices over which the average of distances is computed, it must be given when index_type is either "relative" or "fmi".}

\item{fixed_idx}{the feature indices over which the distance of two constructions is computed, it must be given when index_type is either "relative" or "fmi".}
}
\value{
object of class \code{dist}
}
\description{
Title
}
\examples{
glottosubdata_cnstn <- glottoget(glottodata = "demosubdata_cnstn")
gower_mc_dist <- glottodist_subdata(glottosubdata = glottosubdata_cnstn, metric = "gower", index = "mc")
gower_si_dist <- glottodist_subdata(glottosubdata = glottosubdata_cnstn, metric = "gower", index = "relative", avg_idx = 1:4, fixed_idx = 5:7)
gower_fi_dist <- glottodist_subdata(glottosubdata = glottosubdata_cnstn, metric = "gower", index = "relative", avg_idx = 5:7, fixed_idx = 1:4)
gower_fmi_dist <- glottodist_subdata(glottosubdata = glottosubdata_cnstn_toy, index = "fmi", avg_idx = 1:4, fixed_idx = 5:7)
}
