% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottosimplify.R
\name{glottodrop}
\alias{glottodrop}
\title{Simplify glottodata}
\usage{
glottodrop(glottodata)
}
\arguments{
\item{glottodata}{}
}
\value{

}
\description{
Simplify glottodata
}
\examples{
glottodrop(glottodata)
}
