% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottodrop.R
\name{glottodrop_langtabs}
\alias{glottodrop_langtabs}
\title{Drop language tables from glottodata (keep only meta tables)}
\usage{
glottodrop_langtabs(glottodata)
}
\arguments{
\item{glottodata}{Either glottodata or glottosubdata}
}
\value{

}
\description{
Drop language tables from glottodata (keep only meta tables)
}
