% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottodrop.R
\name{glottodrop_list}
\alias{glottodrop_list}
\title{Drop list structure of glottodata}
\usage{
glottodrop_list(glottodata)
}
\arguments{
\item{glottodata}{glottodata}
}
\value{

}
\description{
Drop list structure of glottodata
}
