% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottodrop.R
\name{glottodrop_meta}
\alias{glottodrop_meta}
\title{Drop metdata tables from glottodata}
\usage{
glottodrop_meta(glottodata)
}
\arguments{
\item{glottodata}{Either glottodata or glottosubdata}
}
\value{

}
\description{
Drop metdata tables from glottodata
}
