% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget.R
\name{glottofiles_cachedir}
\alias{glottofiles_cachedir}
\title{Local directory with glottospace package files}
\usage{
glottofiles_cachedir()
}
\value{

}
\description{
Loads path to local directory where glottospace files are stored and creates directory in a platform independent way in case it doesn't exist.
}
