% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget.R
\name{glottofiles_makedir}
\alias{glottofiles_makedir}
\title{Create local directory for a dirname}
\usage{
glottofiles_makedir(dirname)
}
\arguments{
\item{dirname}{}
}
\value{

}
\description{
Create local directory for a dirname
}
