% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget.R
\name{glottofiles_makepath}
\alias{glottofiles_makepath}
\title{Create local path for a filename}
\usage{
glottofiles_makepath(filename)
}
\arguments{
\item{filename}{}
}
\value{

}
\description{
Create local path for a filename
}
