% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottofilter.R
\name{glottofilter_bycol}
\alias{glottofilter_bycol}
\title{glottofilter by column name}
\usage{
glottofilter_bycol(glottodata, colname, select)
}
\arguments{
\item{glottodata}{glottodata table}

\item{colname}{Character with a single column name}

\item{select}{Things to be selected in that column}
}
\value{

}
\description{
This is a low-level function of glottofilter, allowing for negative selections
}
\examples{
glottodata <- glottoget("glottobase")
glottofilter_bycol(glottodata = glottodata, colname = "family", select = "Indo-European")
glottofilter_bycol(glottodata = glottodata, colname = "family", select = -"Indo-European")
glottofilter_bycol(glottodata = glottodata, colname = "country", select = c("Germany", "Netherlands") )
glottofilter_bycol(glottodata = glottodata, colname = "country", select = -c("Germany", "Netherlands") )
}
\keyword{internal}
