% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottofilter.R
\name{glottofilter_flex}
\alias{glottofilter_flex}
\title{Flexible version of glottofilter}
\usage{
glottofilter_flex(glottodata, colname, select)
}
\arguments{
\item{glottodata}{glottodata table}

\item{colname}{Character with a single column name}

\item{select}{Things to be selected in that column}
}
\value{

}
\description{
This is a more flexible alternative to glottofilter, allowing for negative selections.
}
\examples{
glottodata <- glottoget("glottobase")
glottofilter_flex(glottodata = glottodata, colname = "family", select = "Indo-European")
glottofilter_flex(glottodata = glottodata, colname = "family", select = -"Indo-European")
glottofilter_flex(glottodata = glottodata, colname = "country", select = c("Germany", "Netherlands") )
glottofilter_flex(glottodata = glottodata, colname = "country", select = -c("Germany", "Netherlands") )
glottofilter_flex(glottodata = glottodata, colname = "continent", select = -c("South America", "Europe", "Asia", "Oceania", "Africa") )
}
\keyword{internal}
