% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget.R
\name{glottoget_glottobase}
\alias{glottoget_glottobase}
\title{Get glottobase reference data}
\usage{
glottoget_glottobase(...)
}
\arguments{
\item{...}{Arguments to glottobooster}
}
\description{
Downloads most recent glottolog data and transforms it. This 'glottobase' is used as reference dataset in several functions.
}
\examples{
glottobase <- glottoget_glottobase()
}
\keyword{internal}
