% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget.R
\name{glottoget_glottolog}
\alias{glottoget_glottolog}
\title{Download glottolog data}
\usage{
glottoget_glottolog(data = NULL)
}
\arguments{
\item{data}{One of \code{"glottolog"} or
\code{"glottolog_cldf"}.}
}
\value{
Either a glot or geoglot object.
}
\description{
Download glottolog data
}
\examples{
glottodata <- glottoget_glottolog()
}
