% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget.R
\name{glottoget_remote}
\alias{glottoget_remote}
\title{Get glottodata from online global databases}
\usage{
glottoget_remote(glottodata = NULL)
}
\arguments{
\item{glottodata}{options are:
"glottobase" - default option, an enhanced version of href{https://glottolog.org/}{glottolog}. See \link{glottologbooster} for details.
"glottolog" - this is the raw data downloaded from \href{https://glottolog.org/}{glottolog}
"glottospace" - a simple dataset with glottocodes and a geometry column. This is a subset of all languages in href{https://glottolog.org/}{glottolog} with spatial coordinates.}
}
\value{

}
\description{
Get glottodata from online sources
}
\examples{
glottoget_remote()
}
\seealso{
Other <glottodata>: 
\code{\link{glottoget_path}()},
\code{\link{glottoget}()},
\code{\link{glottosave}()}
}
\concept{<glottodata>}
