% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget_wals.R
\name{glottoget_wals}
\alias{glottoget_wals}
\title{Get WALS data}
\usage{
glottoget_wals(days = NULL)
}
\value{

}
\description{
This function checks whether most recent version of WALS is locally available. If local version is outdated, the newest version will be downloaded.
}
\examples{
glottoget_wals()
}
\keyword{internal}
