% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget_worldatlas.R
\name{glottoget_worldatlas}
\alias{glottoget_worldatlas}
\title{Get World Atlas data from Zenodo}
\usage{
glottoget_worldatlas(
  version = c("contemporary", "traditional"),
  polygon_set = c("features", "languages", "families"),
  download = FALSE,
  dirpath = NULL
)
}
\arguments{
\item{version}{One of "contemporary" or "traditional".}

\item{polygon_set}{One of "features", "languages", or "families".}

\item{download}{Logical. If TRUE, downloads the data from Zenodo.}

\item{dirpath}{Optional. Path to local data directory.}
}
\value{
An `sf` object with the requested World Atlas polygon data.
}
\description{
This function loads World Atlas polygon data from local path or downloads it via Zenodo.
Version options are "contemporary" or "traditional", and polygon_set options are "features", "languages", or "families".
}
\examples{
\donttest{
glottoget_worldatlas()
}
}
