% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottojoin.R
\name{glottojoin_data}
\alias{glottojoin_data}
\title{Join two glottodata tables.}
\usage{
glottojoin_data(glottodata, with, type = "left", id = NULL, ...)
}
\arguments{
\item{glottodata}{User-provided glottodata}

\item{with}{User-provided glottodata}

\item{id}{Optional, name of column with glottocodes}
}
\value{

}
\description{
Join two glottodata tables.
}
\examples{
glottodatax <- glottoget_path(meta = FALSE)
glottodatay <- glottoget_path(meta = FALSE)
glottojoin_data(glottodatax, glottodatay)
}
\seealso{
glottospace_addcoords glottospace_addcoords
}
\keyword{internal}
