% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottojoin.R
\name{glottojoin_datalist}
\alias{glottojoin_datalist}
\title{Join glottosubdata (a list of glottodata tables for multiple languages) into a single glottodata object}
\usage{
glottojoin_datalist(glottodatalist)
}
\arguments{
\item{glottodatalist}{A list of glottodata objects. Column names across languages should be identical.}
}
\value{
A single glottodata object
}
\description{
Join glottosubdata (a list of glottodata tables for multiple languages) into a single glottodata object
}
\examples{
glottodatalist <- glottoget_glottodata("glottosubdata.xlsx", meta = FALSE)
glottojoin_data(glottodatalist = glottodatalist)
}
