% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottojoin.R
\name{glottojoin_meta}
\alias{glottojoin_meta}
\title{Join glottodata with glottometa}
\usage{
glottojoin_meta(glottodata, glottometa)
}
\arguments{
\item{glottodata}{A glottodata table, or a glottodata list with one table.}

\item{glottometa}{A glottometa table, or a glottometa list}

\item{name}{Optional name(s) of glottometa tables}
}
\value{

}
\description{
Join glottodata with glottometa
}
\examples{
glottodata <- glottoget("demodata", meta = TRUE)
glottometa <- glottodata[-1]
glottodata <- glottodata[[1]]
glottojoin_meta(glottodata, glottometa)
}
\keyword{internal}
