% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottojoin.R
\name{glottojoin_space}
\alias{glottojoin_space}
\title{Add coordinates to glottodata.}
\usage{
glottojoin_space(glottodata, id = NULL)
}
\arguments{
\item{glottodata}{User-provided glottodata}

\item{id}{Optional, name of column with glottocodes}
}
\value{

}
\description{
Join glottodata with glottospace (keeps all rows from glottodata). It is recommended to use the function glottospace().
}
\examples{
glottodata <- glottoget_path(meta = FALSE)
glottojoin_space(glottodata)
}
\seealso{
glottospace_addcoords glottospace_addcoords
}
\keyword{internal}
