% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottojoin.R
\name{glottojoin_subdata}
\alias{glottojoin_subdata}
\title{Join glottosubdata (a list of glottodata tables for multiple languages) into a single glottodata object}
\usage{
glottojoin_subdata(glottosubdata)
}
\arguments{
\item{glottosubdata}{A list of glottodata objects. Column names across languages should be identical.}
}
\value{
A single glottodata object
}
\description{
Join glottosubdata (a list of glottodata tables for multiple languages) into a single glottodata object
}
\examples{
glottosubdata <- glottocreate_demosubdata()
glottojoin_data(glottosubdata = glottosubdata)
}
\keyword{internal}
