% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glottolog}
\alias{glottolog}
\title{This is an internally stored version of glottolog data. Use glottoget("glottolog", download = TRUE) to download the latest version}
\format{
A data frame based on glottolog
}
\source{
\url{https://glottolog.org/
}
}
\usage{
glottolog
}
\description{
This is an internally stored version of glottolog data. Use glottoget("glottolog", download = TRUE) to download the latest version
}
\keyword{internal}
\note{version: v4.5


Hammarström, Harald & Forkel, Robert & Haspelmath, Martin & Bank, Sebastian. 2021. Glottolog 4.5. Leipzig: Max Planck Institute for Evolutionary Anthropology. (Available online at https://glottolog.org)

##------ Tue Dec  2 22:50:31 2025 ------##}
