% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottologbooster.R
\name{glottolog_addfamilysizerank}
\alias{glottolog_addfamilysizerank}
\title{Create rank order of family size}
\usage{
glottolog_addfamilysizerank(glottologdata)
}
\arguments{
\item{glottologdata}{}
}
\value{

}
\description{
Create rank order of family size
}
