% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget.R
\name{glottolog_date_local}
\alias{glottolog_date_local}
\title{Check how long ago glottolog was downloaded}
\usage{
glottolog_date_local()
}
\value{
Number of days passed since glottolog data was downloaded for the last time
}
\description{
Check how long ago glottolog was downloaded
}
\keyword{internal}
