% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget.R
\name{glottolog_download}
\alias{glottolog_download}
\title{Download glottolog data}
\usage{
glottolog_download()
}
\description{
This function tries to download glottolog data from zenodo in cldf format.
}
\keyword{internal}
