% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget.R
\name{glottolog_download_cldf}
\alias{glottolog_download_cldf}
\title{Download glottolog data from zenodo, and select relevant data from cldf data}
\usage{
glottolog_download_cldf()
}
\value{

}
\description{
Download glottolog data from zenodo, and select relevant data from cldf data
}
\keyword{internal}
