% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget.R
\name{glottolog_download_webpage}
\alias{glottolog_download_webpage}
\title{Download glottolog data from glottolog website}
\usage{
glottolog_download_webpage()
}
\value{

}
\description{
Download glottolog data from glottolog website
}
\keyword{internal}
