% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget.R
\name{glottolog_download_zenodo}
\alias{glottolog_download_zenodo}
\title{Download most recent version of glottolog from zenodo (cldf format)}
\usage{
glottolog_download_zenodo()
}
\description{
Download most recent version of glottolog from zenodo (cldf format)
}
\keyword{internal}
