% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget.R
\name{glottolog_loadlocal}
\alias{glottolog_loadlocal}
\title{Load locally stored glottolog data}
\usage{
glottolog_loadlocal()
}
\value{

}
\description{
Load locally stored glottolog data
}
\keyword{internal}
