% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottologbooster.R
\name{glottolog_rmbookkeeping}
\alias{glottolog_rmbookkeeping}
\title{Glottolog: remove bookkeeping}
\usage{
glottolog_rmbookkeeping(glottologdata)
}
\arguments{
\item{glottologdata}{data from \href{https://glottolog.org/}{glottolog}, can be downloaded with \code{\link{get_glottolog()}}}
}
\value{

}
\description{
Glottolog: remove bookkeeping
}
\seealso{
Other <glottolog>: 
\code{\link{glottolog_addfamilyname}()},
\code{\link{glottolog_addfamilysizerank}()},
\code{\link{glottolog_addfamilysize}()},
\code{\link{glottolog_addisolates}()},
\code{\link{glottolog_rmartifam}()},
\code{\link{glottolog_rmdialects}()},
\code{\link{glottolog_rmfamilies}()},
\code{\link{glottolog_rmsignlangs}()}
}
\concept{<glottolog>}
\keyword{internal}
