% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottobooster.R
\name{glottolog_rmdialects}
\alias{glottolog_rmdialects}
\title{Glottolog: remove dialects}
\usage{
glottolog_rmdialects(glottologdata)
}
\arguments{
\item{glottologdata}{data from \href{https://glottolog.org/}{glottolog}, can be downloaded with \code{\link{glottoget_glottolog()}}}
}
\value{

}
\description{
Glottolog: remove dialects
}
\seealso{
Other <glottolog>: 
\code{\link{glottolog_L1only}()},
\code{\link{glottolog_addfamilyname}()},
\code{\link{glottolog_addfamilysizerank}()},
\code{\link{glottolog_addfamilysize}()},
\code{\link{glottolog_addisolates}()},
\code{\link{glottolog_rmartifam}()},
\code{\link{glottolog_rmbookkeeping}()},
\code{\link{glottolog_rmfamilies}()},
\code{\link{glottolog_rmsignlangs}()}
}
\concept{<glottolog>}
\keyword{internal}
