% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget.R
\name{glottolog_version_local}
\alias{glottolog_version_local}
\title{Check which version of glottolog is available on your computer}
\usage{
glottolog_version_local()
}
\description{
Check which version of glottolog is available on your computer
}
\keyword{internal}
