% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget.R
\name{glottolog_version_localdatereset}
\alias{glottolog_version_localdatereset}
\title{Reset last modified date of glottolog}
\usage{
glottolog_version_localdatereset()
}
\value{

}
\description{
Reset last modified date of glottolog
}
\keyword{internal}
