% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget.R
\name{glottolog_version_remote}
\alias{glottolog_version_remote}
\title{Check what's the most recent version of glottolog}
\usage{
glottolog_version_remote()
}
\value{

}
\description{
Check what's the most recent version of glottolog
}
\keyword{internal}
