% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottomap.R
\name{glottomap}
\alias{glottomap}
\title{glottomap}
\usage{
glottomap(glottodata, color = NULL, label = NULL, type = NULL, ptsize = NULL)
}
\arguments{
\item{glottodata}{User-provided glottodata}

\item{color}{column name or index to be used to color features (optional)}

\item{label}{Column name or index to be used to label features (optional)}

\item{type}{One of: "static", "dynamic", or "interactive". Defaults to "static" if nothing is provided.}
}
\value{

}
\description{
Create dynamic, static and interactive maps from glottodata
}
\examples{
glottodata <- glottofilter(country = c("Netherlands", "Germany", "Belgium") )
glottodata$country <- as.factor(glottodata$country)
colorpal <- rainbow(length(unique(glottodata$country) ))
glottodata$highlight <- colorpal[glottodata$country]
glottomap(glottodata, color = "highlight")

glottopoints <- glottofilter(continent = "South America")
glottopols <- points2pols(glottopoints, method = "voronoi", continent = "South America")
glottomap(glottodata = glottopols, color = "family_size_rank")
}
