% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottomap.R
\name{glottomap_dynamic}
\alias{glottomap_dynamic}
\title{glottomap_dynamic}
\usage{
glottomap_dynamic(glottodata, label, colorby)
}
\arguments{
\item{glottodata}{User-provided glottodata}

\item{label}{Column name or index to be used to label features (optional)}

\item{colorby}{column name or index to be used to color features (optional)}
}
\value{

}
\description{
glottomap_dynamic
}
\examples{
glottomap_dynamic(glottodata)
}
\keyword{internal}
