% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottomap.R
\name{glottomap_geodata}
\alias{glottomap_geodata}
\title{glottomap_geodata}
\usage{
glottomap_geodata(geodata)
}
\arguments{
\item{geodata}{}
}
\value{

}
\description{
glottomap_geodata
}
\examples{
glottomap_geodata(geodata)
}
\keyword{internal}
