% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottomap.R
\name{glottomap_persist_diagram}
\alias{glottomap_persist_diagram}
\title{Title}
\usage{
glottomap_persist_diagram(glottodata, maxscale)
}
\arguments{
\item{glottodata}{a glottodata is an object of sf with geometry type as `POINT`}

\item{maxscale}{a numeric number, maximum value of the rips filtration, the default unit is "100km"}
}
\value{
a ggplot2 map
}
\description{
Title
}
\examples{
glottopoints <- glottofilter(continent = "South America")
awk <- glottopoints[glottopoints$family == "Arawakan", ]
glottomap_persist_diagram(awk, maxscale = 15)
}
