% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottomap.R
\name{glottomap_static_pacific}
\alias{glottomap_static_pacific}
\title{Create a static (Pacific-centered) map with glottodata}
\usage{
glottomap_static_pacific(
  glottodata,
  color = NULL,
  rivers = FALSE,
  ptsize = NULL
)
}
\arguments{
\item{glottodata}{}
}
\description{
This function returns a static map with glottodata. Data is projected using the Robinson projection centered at the Pacific.
}
\examples{
glottodata <- glottofilter(location = "Australia")
glottomap_static_pacific(glottodata)
}
\keyword{internal}
