% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoplot.R
\name{glottonmds}
\alias{glottonmds}
\title{Nonmetric Multidimensional Scaling}
\usage{
glottonmds(glottodist, k = 2, rm.na = FALSE)
}
\arguments{
\item{k}{Number of dimensions}

\item{rm.na}{Whether NAs should be removed (default is FALSE)}

\item{dist}{dist object or distance matrix}
}
\description{
Nonmetric Multidimensional Scaling
}
\examples{
glottodata <- glottoget("demodata", meta = TRUE)
glottodist <- glottodist(glottodata = glottodata)
glottonmds <- glottonmds(glottodist = glottodist, k = 2)
}
\seealso{
Other <glottoplot>: 
\code{\link{glottonmds_scores}()}
}
\concept{<glottoplot>}
\keyword{internal}
