% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoplot.R
\name{glottonmds_scores}
\alias{glottonmds_scores}
\title{Get nmds scores}
\usage{
glottonmds_scores(glottonmds)
}
\arguments{
\item{glottonmds}{a glottonmnds object created with glottonmds_run()}
}
\value{

}
\description{
Obtain nmds scores of a glottonmds object
}
\examples{
glottonmds_scores(glottonmds)
}
\seealso{
Other <glottoplot>: 
\code{\link{glottonmds}()}
}
\concept{<glottoplot>}
\keyword{internal}
