% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoplot.R
\name{glottoplot}
\alias{glottoplot}
\title{Visualize distances}
\usage{
glottoplot(glottodist, type = NULL, filename = NULL, ...)
}
\arguments{
\item{glottodist}{A dist object created with \code{\link{glottodist}}}

\item{type}{The type of plot: "heatmap", "nmds"}

\item{filename}{Optional filename if output should be saved.}

\item{...}{Additional arguments passed to plotnmds (if type = "nmds").}
}
\value{

}
\description{
This function offers different types of visualizations to show linguistic distances.
}
\examples{
glottodata <- glottoget_path(meta = TRUE)
glottodist <- glottodist(glottodata = glottodata)
}
