% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoplot.R
\name{glottoplot_heatmap}
\alias{glottoplot_heatmap}
\title{Plot a heatmap of a distance matrix}
\usage{
glottoplot_heatmap(glottodist, filename = NULL)
}
\arguments{
\item{glottodist}{A dist object created with \code{\link{glottodist}}}

\item{filename}{Optional filename if output should be saved.}
}
\value{

}
\description{
Plot a heatmap of a distance matrix
}
\examples{
glottoplot_heatmap(glottodist)
}
