% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoplot.R
\name{glottoplot_naviewer}
\alias{glottoplot_naviewer}
\title{Show data coverage (view NAs)}
\usage{
glottoplot_naviewer(data, id = NULL, rm.na = TRUE)
}
\arguments{
\item{data}{Any dataset}

\item{id}{column name with IDs}

\item{rm.na}{Whether rows without id should be removed.}
}
\value{

}
\description{
This function plots the NAs in a dataset. If you used another coding to
specify missing data, you should run \code{cleandata_recodemissing} or \code{cleanglottodata} first. If you'd
like some more advanced ways of handling NAs, you might check out the
\code{naniar} package.
}
\examples{
glottodata <- glottoget("demodata", meta = TRUE)
glottodata <- glottoclean(glottodata = glottodata)
data <- glottodata[[1]]
glottoplot_naviewer(data, id = "glottocode")
}
\keyword{internal}
