% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoplot.R
\name{glottoplot_nmds}
\alias{glottoplot_nmds}
\title{Plot nmds in 3d}
\usage{
glottoplot_nmds(
  nmds,
  scoresdata,
  color = NULL,
  ptsize = NULL,
  label = NULL,
  filename = NULL
)
}
\arguments{
\item{nmds}{An nmds object}

\item{scoresdata}{scoresdata}

\item{color}{color}

\item{ptsize}{ptsize}

\item{label}{label}

\item{filename}{optional filename if output should be saved.}
}
\description{
Plot nmds in 3d
}
\examples{
glottodata <- glottoget("demodata", meta = TRUE)
glottodist <- glottodist(glottodata = glottodata)
nmds <- glottonmds(glottodist, k = 2)
scores <- glottonmds_scores(nmds)
scoresdata <- glottojoin_base(scores)

glottoplot_nmds(nmds = nmds, scoresdata = scoresdata, color = "family", ptsize = "isolate")
glottoplot_nmds(nmds = nmds, scoresdata = scoresdata, color = "isolate")
}
\keyword{internal}
