% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoclean.R
\name{glottorecode}
\alias{glottorecode}
\title{Recode values across a glottodataset}
\usage{
glottorecode(glottodata, tofalse = NULL, totrue = NULL, tona = NULL)
}
\arguments{
\item{glottodata}{glottodata list}

\item{tofalse}{Values to recode to FALSE}

\item{totrue}{Values to recode to TRUE}

\item{tona}{Values to recode to NA}
}
\description{
This function recodes values within a glottodataset to NA or TRUE/FALSE. Recoding is done based on column types in the structure table. Run glottocreate_structuretable() to create one.
}
\examples{
glottodata <- glottoget("demodata", meta = TRUE)
glottodata <- glottorecode(glottodata, tona = c("?", "missing"))
}
\concept{<glottorecode><glottoclean>}
\keyword{internal}
