% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoclean.R
\name{glottorecode_logical}
\alias{glottorecode_logical}
\title{Recode character columns to TRUE/FALSE}
\usage{
glottorecode_logical(glottodata, totrue = NULL, tofalse = NULL)
}
\arguments{
\item{glottodata}{glottodata list}
}
\description{
Recode character columns to TRUE/FALSE
}
\examples{
glottorecode_logical(glottodata, totrue = c("y", "Y", 1), tofalse = c("n", "N", 0))
}
\keyword{internal}
