% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoclean.R
\name{glottorecode_missing}
\alias{glottorecode_missing}
\title{Recode missing values to NA}
\usage{
glottorecode_missing(glottodata, tona)
}
\arguments{
\item{glottodata}{User-provided glottodata}

\item{tona}{Optional, additional values to recode to NA}
}
\value{

}
\description{
Recode missing values to NA
}
\examples{
glottodata <- glottoget("demodata", meta = TRUE)
glottorecode_missing(glottodata, tona = "?")
}
\concept{<glottorecode>}
\keyword{internal}
