% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottosave.R
\name{glottosave}
\alias{glottosave}
\title{Save glottodata, maps and plots}
\usage{
glottosave(glottodata, filename = NULL)
}
\arguments{
\item{glottodata}{User-provided glottodata}

\item{filename}{Filename either with or without file extension}
}
\value{

}
\description{
If no filename is provided, the name of the glottodata object will be used.
}
\details{
If no file extension is provided, a sensible default file extension is chosen. Dynamic maps
(tmap) are saved in .html format, static maps (tmap) are saved as .png.
Spatial data (sf) are saved as geopackage (.GPKG) by default, but .shp is
also possible.
}
\examples{

glottodata <- glottoget("demodata", meta = FALSE)
Saves as .xlsx
glottosave(glottodata, filename = "glottodata")

glottospacedata <- glottospace(glottodata)
Saves as .GPKG
glottosave(glottodata, filename = "glottodata")

glottomap <- glottomap(glottodata)
Saves as .png
glottosave(glottomap, filename = "glottomap")

Saves as .html
glottomap <- glottomap(glottodata, type = "dynamic")
}
\seealso{
glottoget_glottodata

Other <glottodata>: 
\code{\link{glottoget_path}()},
\code{\link{glottoget_remote}()},
\code{\link{glottoget}()}
}
\concept{<glottodata>}
