% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottosearch.R
\name{glottosearch}
\alias{glottosearch}
\title{Search within glottodata for languages, glottocodes, etc.}
\usage{
glottosearch(glottodata = NULL, find, partialmatch = TRUE, columns = NULL)
}
\arguments{
\item{glottodata}{Any linguistic or cultural dataset. Default is to search within glottolog.}

\item{find}{Character string to search for, this can be the name of a language, a family, a dialect, a glottocode, isocode.}

\item{partialmatch}{By default, partial matches will be returned as well. In case you only want exact matches, this argument should be set to FALSE.}

\item{columns}{By default, the entire dataset is searched, but optionally the search can be limited to specific columns.}
}
\value{

}
\description{
Search within glottodata for languages, glottocodes, etc.
}
\examples{
glottosearch(find = "Yucuni")
glottosearch(find = "Yucuni", columns = "name")
}
