% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottosimplify.R
\name{glottosimplify_droplist}
\alias{glottosimplify_droplist}
\title{Drop list structure of glottodata}
\usage{
glottosimplify_droplist(glottodata)
}
\arguments{
\item{glottodata}{glottodatalist}
}
\value{

}
\description{
Select glottodatatable from a glottodatalist
}
\keyword{internal}
