% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottosimplify.R
\name{glottosimplify_dropmeta}
\alias{glottosimplify_dropmeta}
\title{Drop metadata tables from glottodata}
\usage{
glottosimplify_dropmeta(glottodata)
}
\arguments{
\item{glottodata}{Either glottodata or glottosubdata}
}
\value{

}
\description{
Drop metadata tables from glottodata
}
