% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottosimplify.R
\name{glottosimplify_dropspatial}
\alias{glottosimplify_dropspatial}
\title{Drop spatial properties from glottodata}
\usage{
glottosimplify_dropspatial(glottodata)
}
\arguments{
\item{glottodata}{Either glottodata or glottosubdata}
}
\value{

}
\description{
Drop spatial properties from glottodata
}
