% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottosimplify.R
\name{glottosimplify_dropunits}
\alias{glottosimplify_dropunits}
\title{Drop units from glottodata}
\usage{
glottosimplify_dropunits(glottodata)
}
\arguments{
\item{glottodata}{Either glottodata or glottosubdata}
}
\description{
Drop units from glottodata
}
\keyword{internal}
