% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottosimplify.R
\name{glottosimplify_langtabs}
\alias{glottosimplify_langtabs}
\title{Drop language tables from glottodata (keep only meta tables)}
\usage{
glottosimplify_langtabs(glottodata)
}
\arguments{
\item{glottodata}{Either glottodata or glottosubdata}
}
\description{
Drop language tables from glottodata (keep only meta tables)
}
\keyword{internal}
