% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottosimplify.R
\name{glottosimplify_list2tab}
\alias{glottosimplify_list2tab}
\title{Select glottodatatable from a glottodatalist}
\usage{
glottosimplify_list2tab(glottodata)
}
\arguments{
\item{glottodata}{glottodatalist}
}
\value{

}
\description{
Select glottodatatable from a glottodatalist
}
