% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottosimplify.R
\name{glottosimplify_submerge}
\alias{glottosimplify_submerge}
\title{Conditionally merge tables from glottosubdata}
\usage{
glottosimplify_submerge(glottosubdata)
}
\arguments{
\item{glottosubdata}{glottosubdata.}
}
\value{

}
\description{
Check whether input is glottosubdata and merge tables if TRUE.
}
\keyword{internal}
