% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottospace.R
\name{glottospace}
\alias{glottospace}
\title{Make glottodata spatial}
\usage{
glottospace(glottodata, glottospace)
}
\arguments{
\item{glottodata}{A glottodata table, or list of glottodata tables}

\item{glottospace}{By default, glottopoints are returned, glottopols are also supported.}
}
\value{

}
\description{
This function takes glottodata (either a table or list of tables) and turns it into spatial points or polygons.
}
\examples{
glottodata <- glottoget()
glottodataspace <- glottospace(glottodata)
}
