% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottospace.R
\name{glottospace_addcoords}
\alias{glottospace_addcoords}
\title{Make glottodata spatial}
\usage{
glottospace_addcoords(glottodata)
}
\arguments{
\item{glottodata}{User-provided glottodata}
}
\value{
glottodata with a GEOMETRY column
}
\description{
Adds coordinates from glottolog to user-provided glottodata.
If glottodata does not contain a geometry column, it will be added.
}
\examples{
glottodata <- glottoget("demodata")
glottospace_addcoords(glottodata)
}
\keyword{internal}
