% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottospace.R
\name{glottospace_addcountries}
\alias{glottospace_addcountries}
\alias{glottodata_addcontinents}
\alias{glottodata_addregions}
\title{Add names of countries, continents, and regions to glottodata}
\usage{
glottospace_addcountries(glottodata)
}
\arguments{
\item{glottodata}{User-provided glottodata}
}
\value{
A spatial (sf) object
}
\description{
Spatial intersection between glottodata and
\href{https://www.naturalearthdata.com/}{naturalearthdata} and extracts names
of countries, continents, and regions.
}
\examples{
glottodata <- glottoget("demodata")
glottospace_addcountries(glottodata)
}
\keyword{internal}
