% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottospace.R
\name{glottospace_coords2sf}
\alias{glottospace_coords2sf}
\title{Convert glottodata with lat/lon columns to simple feature class}
\usage{
glottospace_coords2sf(glottodata, lon = "longitude", lat = "latitude")
}
\arguments{
\item{glottodata}{glottodata table}

\item{lon}{Column name containing longitude}

\item{lat}{Column name containing latitude}
}
\value{

}
\description{
This function is used by glottobooster().
}
\keyword{internal}
