% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottosplit.R
\name{glottosplit}
\alias{glottosplit}
\title{Split metadata and data from glottodata or glottosubdata}
\usage{
glottosplit(glottodata)
}
\arguments{
\item{glottodata}{glottodata or glottosubdata (either with or without metadata)}
}
\value{
A list of length 2
}
\description{
Splits data and metadata from glottodata or glottosubdata and returns a list of two (1: glottodata tables, and 2: metadata tables )
}
\details{
If data does not contain metadata, the second list element will be NA.
}
\examples{
glottodata <- glottoget("demodata", meta = TRUE)
splitted <- glottosplit(glottodata)
merged <- glottojoin(splitted[[1]], splitted[[2]])
}
\keyword{internal}
