% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottostat.R
\name{glottostat_dist_permanova}
\alias{glottostat_dist_permanova}
\title{Permanova across all groups (overall or pairwise)}
\usage{
glottostat_dist_permanova(
  glottodist = NULL,
  glottodata = NULL,
  comparison = NULL,
  sample = NULL,
  permutations = NULL,
  by = NULL
)
}
\arguments{
\item{glottodist}{a dist object}

\item{glottodata}{glottodata contains sample}

\item{comparison}{Either "overall" or "pairwise"}

\item{sample}{sample table (optional). By default, searches for sample table in glottodata/glottosubdata.}

\item{permutations}{Number of permutations (default is 999)}

\item{by}{the column name of "sample", over which to compute the permanova.}
}
\description{
This function takes a dist object and performs a
Permutational Multivariate Analysis of Variance (PERMANOVA). It can be used
to test whether two or more groups are significantly different from each
other (by specifying the \code{comparison} argument with either 'overall' or
'pairwise').
}
\details{
The argument \code{by} is the name of a column in the sample table,
which can be either provided by a "sample" sheet in \code{glottodata} or given by the argument \code{sample}.
The default value of \code{by} is "group".
The function uses \code{by} to do the comparisons.
The function calls \code{vegan::adonis2()}, type \code{?adonis2} for more details.
}
\examples{
glottodata <- glottoget("demodata", meta = TRUE)
glottodist <- glottodist(glottodata, metric = "gower")
glottostat_dist_permanova(glottodist = glottodist, glottodata = glottodata, comparison = "pairwise")


}
