% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottostat.R
\name{glottostat_dist_permanova_mci}
\alias{glottostat_dist_permanova_mci}
\title{Title}
\usage{
glottostat_dist_permanova_mci(
  glottodist = NULL,
  glottodata = NULL,
  comparison = NULL,
  sample = NULL,
  permutations = NULL,
  by = NULL
)
}
\arguments{
\item{by}{}
}
\description{
Title
}
