% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_functions.R
\name{gs_condistagg}
\alias{gs_condistagg}
\title{Aggregate distances between constructions per language
#' @param aggregate One of c('mean', 'min', 'sum') indicating how distances should be aggregated to language level.('min' returns best match)}
\usage{
gs_condistagg(condist, glottocodes, aggregation)
}
\description{
Aggregate distances between constructions per language
#' @param aggregate One of c('mean', 'min', 'sum') indicating how distances should be aggregated to language level.('min' returns best match)
}
