% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottodist.R
\name{gs_condistconvert}
\alias{gs_condistconvert}
\title{Convert distances between constructions per language}
\usage{
gs_condistconvert(
  condist,
  glottocodes,
  groups = NULL,
  thresval = NULL,
  threstype = "absolute"
)
}
\arguments{
\item{groups}{Vector of two groups (e.g. language families) for which average should be calculated.
Default is to use all pairwise distances to calculate average.}
}
\description{
Convert distances between constructions per language
}
