% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gs_emptydistmat}
\alias{gs_emptydistmat}
\title{Create empty distance matrix}
\usage{
gs_emptydistmat(names)
}
\arguments{
\item{names}{}
}
\value{

}
\description{
Create empty distance matrix
}
\examples{
gs_emptydistmat(names = glottocodes)
}
