% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joindata.R
\name{join_glottobase}
\alias{join_glottobase}
\title{Join glottodata with glottobase}
\usage{
join_glottobase(glottodata, id = NULL)
}
\arguments{
\item{id}{}
}
\value{

}
\description{
Keeps all rows from glottodata. Can be used to add features from glottobase.
}
