% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joindata.R
\name{join_glottodata}
\alias{join_glottodata}
\title{Join glottodata for multiple languages into a single glottodata object}
\usage{
join_glottodata(langlist)
}
\arguments{
\item{langlist}{A list of glottodata objects. Column names across languages should be identical.}
}
\value{
A single glottodata object
}
\description{
Join glottodata for multiple languages into a single glottodata object
}
\examples{
glottodata <- get_glottodata("glottosubdata.xlsx", meta = FALSE)
join_glottodata(langlist = glottodata)
}
