% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottojoin.R
\name{join_glottodatalist}
\alias{join_glottodatalist}
\title{Join glottosubdata (a list of glottodata tables for multiple languages) into a single glottodata object}
\usage{
join_glottodatalist(glottodatalist)
}
\arguments{
\item{glottodatalist}{A list of glottodata objects. Column names across languages should be identical.}
}
\value{
A single glottodata object
}
\description{
Join glottosubdata (a list of glottodata tables for multiple languages) into a single glottodata object
}
\examples{
glottodatalist <- glottoget_glottodata("glottosubdata.xlsx", meta = FALSE)
join_glottodata(glottodatalist = glottodatalist)
}
