% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joindata.R
\name{join_glottodist}
\alias{join_glottodist}
\title{Join a dataset with a dist object}
\usage{
join_glottodist(data, id = "glottocode", dist, rm.na = TRUE)
}
\arguments{
\item{data}{A data.frame}

\item{id}{Column with IDs in \code{data} that match names of dist object. By default, the "glottocode" column is used.}

\item{dist}{A dist object}

\item{rm.na}{Default is to remove NAs.}
}
\value{
Data frame
}
\description{
Join a dataset with a dist object
}
\examples{
path <- "C:/Users/sjnor/surfdrive/Projecten en schrijfsels/Papers in progress/Isolates/output/south america all.gpkg"
sa <- sf::st_read(dsn = path)
dist <- pointdist_bird(sa, label = "glottocode")
distdf <- joindatadist(data = sa, idcol = "glottocode", dist = dist)

After joining, you can subset the distance columns by using the IDs:
distdf[, distdf$glottocode]
}
