% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottojoin.R
\name{join_glottometa}
\alias{join_glottometa}
\title{Join glottodata with glottometa}
\usage{
join_glottometa(glottodata, glottometa, name = NULL)
}
\arguments{
\item{glottodata}{A glottodata table, or a glottodata list with one table.}

\item{glottometa}{A glottometa table, or a glottometa list}

\item{name}{}
}
\value{

}
\description{
Join glottodata with glottometa
}
\examples{
glottodata <- glottoget_path(meta = TRUE)
glottometa <- glottodata[-1]
glottodata <- glottodata[[1]]
join_glottometa(glottodata, glottometa)
}
