% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joindata.R
\name{join_glottospace}
\alias{join_glottospace}
\title{Add coordinates to glottodata.}
\usage{
join_glottospace(glottodata, id = NULL)
}
\arguments{
\item{id}{}
}
\value{

}
\description{
Join glottodata with glottospace (keeps all rows from glottodata)
}
