% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodist.R
\name{meandist}
\alias{meandist}
\title{Calculate mean distance from one features to all other features}
\usage{
meandist(dist)
}
\arguments{
\item{dist}{\code{dist} object}
}
\value{

}
\description{
Calculate mean distance from one features to all other features
}
