% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_geo.R
\name{mergevec}
\alias{mergevec}
\title{Merge multiple shapefiles}
\usage{
mergevec(paths = NULL, selection = NULL, outfile = NULL, overwrite = FALSE)
}
\arguments{
\item{overwrite}{}
}
\value{

}
\description{
Merge multiple shapefiles
}
\keyword{internal}
