% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_geo.R
\name{mosaicras}
\alias{mosaicras}
\title{Mosaic raster tiles}
\usage{
mosaicras(
  path = NULL,
  outfile = "mosaic.tif",
  overwrite = FALSE,
  na.rm = TRUE,
  subdirs = TRUE
)
}
\arguments{
\item{path}{path to folder containing multiple raster files}

\item{outfile}{file path where to save resulting raster mosaic}

\item{subdirs}{}
}
\value{

}
\description{
Mosaic raster tiles
}
\keyword{internal}
