% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkglottodata.R
\name{naviewer}
\alias{naviewer}
\title{Show data coverage (view NAs)}
\usage{
naviewer(data, id)
}
\arguments{
\item{id}{}
}
\value{

}
\description{
This function plots the NAs in a dataset. If you used another coding to specify missing data, you should run \code{cleanglottodata} first.
}
