% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodist.R
\name{nearestdistance}
\alias{nearestdistance}
\title{Get distance to nearest features}
\usage{
nearestdistance(dist, n = 1)
}
\arguments{
\item{dist}{dist object to obtain distances from}

\item{n}{The n nearest features for which distances should be obtained, default (n = 1) is only distance to nearest neighbor.}
}
\value{
numeric vector with distances if only nearest distance is requested (n = 1) and a data.frame otherwise.
}
\description{
Get distance to nearest features
}
\examples{
nearestdistance(dist)
}
