% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodist.R
\name{nearestline_bird}
\alias{nearestline_bird}
\title{Calculate distance from each point to nearest line, as the bird flies (great circle distance)}
\usage{
nearestline_bird(points, lines, label)
}
\arguments{
\item{points}{Spatial point object}

\item{lines}{Spatial line object}

\item{label}{Column name or index to be used for labeling objects.}
}
\value{
dist object with distances from each point (km) to the nearest line.
}
\description{
Great circle distances (Haversine distances) are calculated between points an nearest line (e.g. river, road, etc. ).
}
\examples{
nearestline_bird(points = points, lines = rivers, label = "glottocode")
nearestline_bird(points = points, lines = roads, label = "glottocode")
}
