% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdistance_nmds.R
\name{nmds}
\alias{nmds}
\title{Nonmetric Multidimensional Scaling}
\usage{
nmds(dist, k = 2, rm.na = FALSE)
}
\arguments{
\item{rm.na}{}
}
\value{

}
\description{
Nonmetric Multidimensional Scaling
}
\examples{
dist <- glottodist(glottodata = glottodata, structure = structure)
nmds <- nmds(dist = dist, k = 2)
}
