% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_general.R
\name{nonna}
\alias{nonna}
\title{Select non-NA elements from vector}
\usage{
nonna(v, max1 = FALSE)
}
\arguments{
\item{v}{}
}
\description{
Select non-NA elements from vector. If all elements are NA, NA will be returned.
}
