% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phoible}
\alias{phoible}
\title{This is an internally stored version of PHOIBLE data. Use glottoget("phoible", download = TRUE) to download the latest version}
\format{
A data frame based on PHOIBLE
}
\source{
\url{https://phoible.org/
}
}
\usage{
phoible
}
\description{
This is an internally stored version of PHOIBLE data. Use glottoget("phoible", download = TRUE) to download the latest version
}
\keyword{internal}
\note{version: v2.0.1


The data of PHOIBLE as CLDF StructureDataset.

Cite as:


Moran, Steven & McCloy, Daniel (eds.) 2019. PHOIBLE. Jena: Max Planck Institute for the Science of Human History. (Available online at https://phoible.org)


##------ Mon Feb 26 23:32:21 2024 ------##}
