% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phoible_raw}
\alias{phoible_raw}
\title{This is an internally stored version of raw PHOIBLE data. Use glottoget("phoible_raw", download = TRUE) to download the latest version}
\format{
A data frame based on PHOIBLE
}
\source{
\url{https://phoible.org/
}
}
\usage{
phoible_raw
}
\description{
This is an internally stored version of raw PHOIBLE data. Use glottoget("phoible_raw", download = TRUE) to download the latest version
}
\keyword{internal}
\note{version: v2.0.1


The data of PHOIBLE as CLDF StructureDataset.

Cite as:


Moran, Steven & McCloy, Daniel (eds.) 2019. PHOIBLE. Jena: Max Planck Institute for the Science of Human History. (Available online at https://phoible.org)


##------ Thu Nov 13 17:57:33 2025 ------##}
