% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget_phoible.R
\name{phoible_transpose}
\alias{phoible_transpose}
\title{Title}
\usage{
phoible_transpose(phoible_data)
}
\arguments{
\item{phoible_data}{A non-spatial phoible dataset}
}
\value{
an sf object
}
\description{
Title
}
\examples{
phoible_sf <- phoible_transpose(glottospace::phoible_raw)
}
