% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe}
\arguments{
\item{lhs, rhs}{specify what lhs and rhs are}
}
\description{
Put description here
}
\examples{
# some examples if you want to highlight the usage in the package
}
