% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdistance_nmds.R
\name{plotnmds}
\alias{plotnmds}
\title{Plot NMDS}
\usage{
plotnmds(
  nmds,
  scoresdata,
  colorby = NULL,
  sizeby = NULL,
  labelby = NULL,
  filename = NULL
)
}
\arguments{
\item{rm.na}{}
}
\value{

}
\description{
Plot NMDS
}
\examples{
scores <- nmds_scores(nmds)
glottodata <- glottofilter(glottodata = get_glottobase(), glottocode = scores$glottocode)
scoresdata <- join_glottodata(data = scores, glottodata = glottodata)
plotnmds(nmds = nmds, scoresdata = scoresdata, colorby = "family_name", sizeby = "isolate")
}
