% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodist.R
\name{pointdist_topo}
\alias{pointdist_topo}
\title{Distances between spatial points, over land}
\usage{
pointdist_topo(points, label, topography)
}
\arguments{
\item{points}{geoglot or sf object}

\item{label}{Column name or index to be used for labeling objects.}

\item{topography}{raster object or path to raster object}
}
\value{
dist object
}
\description{
Least-cost distance between spatial points. Based on the Dijkstra (1959)
algorithm as implemented in the gdistance package (van Etten, 2017)
}
\examples{
ppath <- "C:/Users/sjnor/surfdrive/Projecten en schrijfsels/Papers in progress/Isolates/output/nwa.gpkg"
topopath <- "D:/Global/Topography/SRTM/250m/South America/SRTM250mSA.tif"
points <- sf::st_read(ppath)
points <- sf::st_geometry(points)
points <- points[c(1:10),] # for testing only
pointdist_topo(points = points, topography = topopath)
}
\seealso{
Other <geodist>: 
\code{\link{geodist}()},
\code{\link{nearestline_bird}()},
\code{\link{pointdist_bird}()}
}
\concept{<geodist>}
