% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcldf.R
\name{read_bib}
\alias{read_bib}
\title{Reads a BibTeX file into a dataframe}
\usage{
read_bib(dir, bib = "sources.bib")
}
\arguments{
\item{dir}{the directory the BibTeX file is in.}

\item{bib}{the name of the BibTeX file (default="sources.bib")}
}
\value{
A tibble dataframe
}
\description{
Reads a BibTeX file into a dataframe
}
