% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcldf.R
\name{read_cldf}
\alias{read_cldf}
\title{Reads a Cross-Linguistic Data Format dataset into an object.}
\usage{
read_cldf(mdpath)
}
\arguments{
\item{mdpath}{the path to the directory or metadata.json file.}
}
\value{
A `cldf` object
}
\description{
Reads a Cross-Linguistic Data Format dataset into an object.
}
\examples{
cldfobj <- cldf(system.file("extdata/huon", "cldf-metadata.json", package = "rcldf"))
}
