% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialdata.R
\name{spatialfilter}
\alias{spatialfilter}
\title{Open geodata (raster or vector format)}
\usage{
spatialfilter(path, name = NULL)
}
\arguments{
\item{path}{Path to raster or vector file}

\item{name}{Optional}
}
\value{

}
\description{
Open geodata (raster or vector format)
}
