% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_general.R
\name{sumna}
\alias{sumna}
\title{Sum of vector elements}
\usage{
sumna(v)
}
\arguments{
\item{v}{}
}
\description{
Calculate sum of all vector elements that are not NA. If all elements are NA, NA will be returned (in contrast to sum).
}
