% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdistance_nmds.R
\name{viewnmds}
\alias{viewnmds}
\title{Title}
\usage{
viewnmds(nmds, filename = NULL, dist = NULL, view = "nmds", rm.na = TRUE)
}
\arguments{
\item{rm.na}{}
}
\value{

}
\description{
Title
}
\examples{
scores <- nmds_scores(nmds)
glottofilter(glottodata = glottobase(), glottocode = scores$glottocode)
}
