% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wals}
\alias{wals}
\title{WALS data}
\format{
A data frame based on The World Atlas of Language Structures
}
\source{
\url{https://wals.info/
}
}
\usage{
wals
}
\description{
This is an internally stored version of WALS data. Use glottoget("wals", download = TRUE) to download the latest version
}
\keyword{internal}
\note{version: v2020.1


Cite the source of the dataset as:


Dryer, Matthew S. & Haspelmath, Martin (eds.) 2013. The World Atlas of Language Structures Online. Leipzig: Max Planck Institute for Evolutionary Anthropology. (Available online at https://wals.info)


##------ Thu Nov 13 17:57:26 2025 ------##}
