% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget_wals.R
\name{wals_date_local}
\alias{wals_date_local}
\title{Check how long ago WALS data was downloaded}
\usage{
wals_date_local()
}
\value{
Number of days passed since WALS data was downloaded for the last time
}
\description{
Check how long ago WALS data was downloaded
}
\keyword{internal}
