% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget_wals.R
\name{wals_download_cldf}
\alias{wals_download_cldf}
\title{Download WALS data from zenodo, and select relevant data from cldf data}
\usage{
wals_download_cldf()
}
\value{

}
\description{
Download WALS data from zenodo, and select relevant data from cldf data
}
\keyword{internal}
