% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget_wals.R
\name{wals_download_zenodo}
\alias{wals_download_zenodo}
\title{Download most recent version of WALS from zenodo (cldf format)}
\usage{
wals_download_zenodo()
}
\value{

}
\description{
Download most recent version of WALS from zenodo (cldf format)
}
