% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget_wals.R
\name{wals_loadlocal_raw}
\alias{wals_loadlocal_raw}
\title{Load locally stored WALS data (without joining with glottolog)}
\usage{
wals_loadlocal_raw()
}
\value{

}
\description{
Load locally stored WALS data (without joining with glottolog)
}
