% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget_wals.R
\name{wals_version_local}
\alias{wals_version_local}
\title{Check which version of WALS is available on your computer}
\usage{
wals_version_local()
}
\value{

}
\description{
Check which version of WALS is available on your computer
}
\keyword{internal}
