% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget_wals.R
\name{wals_version_localdatereset}
\alias{wals_version_localdatereset}
\title{Reset last modified date of glottolog}
\usage{
wals_version_localdatereset()
}
\value{

}
\description{
Reset last modified date of glottolog
}
\keyword{internal}
