% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget_wals.R
\name{wals_version_remote}
\alias{wals_version_remote}
\title{Check what's the most recent version of WALS}
\usage{
wals_version_remote()
}
\value{

}
\description{
Check what's the most recent version of WALS
}
