% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{worldpol}
\alias{worldpol}
\title{This is an internally stored version of political boundaries of the world (obtained from rnaturalearth).}
\format{
An sf object
}
\source{
\url{https://www.naturalearthdata.com/
}
}
\usage{
worldpol
}
\description{
This is an internally stored version of political boundaries of the world (obtained from rnaturalearth).
}
\keyword{internal}
