% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkglottodata.R
\name{checkdata_lscolcount}
\alias{checkdata_lscolcount}
\title{Check whether number of columns are identical across all glottodata objects in a list}
\usage{
checkdata_lscolcount(langlist)
}
\arguments{
\item{langlist}{}
}
\value{

}
\description{
Check whether number of columns are identical across all glottodata objects in a list
}
\keyword{internal}
