% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkglottodata}
\alias{checkglottodata}
\title{Quality check of user-provided glottodata}
\usage{
checkglottodata(data = NULL, idcol = NULL)
}
\arguments{
\item{messages}{}
}
\value{

}
\description{
Go step by step through a user-provided glottodataset. Messages indicate
}
\examples{
dpath <- "C:/Users/sjnor/surfdrive/Shared/SAPPHIRE/Output/Presentations/Isolates_WS-Athens/data/dbase_isolates_V2.xlsx"
data <- get_sheetdata(path = dpath, sheets = 1)
checkglottodata(data = glottodata, idcol = 1)

checkglottodata(data = glottodata, idcol = "glottocode")
lapply(glottodatalist, checkglottodata, idcol = "glottosubcode")
}
