% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocheck.R
\name{checkglottodata}
\alias{checkglottodata}
\title{Quality check of user-provided glottodata}
\usage{
checkglottodata(glottodata, diagnostic = TRUE)
}
\arguments{
\item{glottodata}{User-provided glottodata}

\item{diagnostic}{If TRUE (default) a data viewer will be opened to show the levels of each variable (including NAs), and a data coverage plot will be shown.}
}
\value{

}
\description{
Go through a user-provided glottodataset and check whether:
- one column exists with the name "glottocode"
- there are rows without a glottocode  (missing IDs)
- there are rows with duplicated glottocodes (duplicate IDs)
- all variables have at least two levels
- all glottocodes are valid
}
\examples{
glottodata <- get_glottodata(meta = FALSE)
checkglottodata(glottodata = glottodata)
}
\seealso{
Other <glottocheck>: 
\code{\link{checkglottosubdata}()},
\code{\link{glottocheck_hasmeta}()},
\code{\link{glottocheck}()}
}
\concept{<glottocheck>}
