% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkglottodata.R
\name{checkglottodata}
\alias{checkglottodata}
\title{Quality check of user-provided glottodata}
\usage{
checkglottodata(glottodata, show = TRUE)
}
\arguments{
\item{glottodata}{User-provided glottodata}

\item{show}{If TRUE (default) a data viewer will be opened to show the levels of each variable (including NAs)}

\item{id}{Column index or column name with IDs (glottocodes, glottosubcodes, or glottosubsubcodes). Default is "glottocode".}
}
\value{

}
\description{
Go through a user-provided glottodataset and check:
- No missing IDs
- No duplicate IDs
- All variables have at least two levels
- All glottocodes are valid
}
\examples{
dpath <- "C:/Users/sjnor/surfdrive/Shared/SAPPHIRE/Output/Presentations/Isolates_WS-Athens/data/dbase_isolates_V2.xlsx"
glottodata <- get_sheetdata(path = dpath, sheets = 1)
checkglottodata(glottodata = glottodata, id = 1)

checkglottodata(glottodata = glottodata)
lapply(glottodatalist, checkglottodata)
}
