% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocheck.R
\name{checkglottosubdata}
\alias{checkglottosubdata}
\title{Quality check of user-provided glottosubdata}
\usage{
checkglottosubdata(glottosubdata, diagnostic = TRUE)
}
\arguments{
\item{glottosubdata}{User-provided glottosubdata}

\item{diagnostic}{If TRUE (default) a data viewer will be opened to show the levels of each variable (including NAs), and a data coverage plot will be shown.}
}
\value{

}
\description{
Quality check of user-provided glottosubdata
}
\examples{
glottosubdata <- get_glottodata(meta = FALSE, dummy = "glottosubdata")
glottosubdata <- join_glottodata(glottosubdata)
checkglottosubdata(glottosubdata)

Better to join_glottodata first, instead of the following approach (because checks only within each language for duplicates etc.)
lapply(glottosubdata, checkglottosubdata)
}
\seealso{
Other <glottocheck>: 
\code{\link{checkglottodata}()},
\code{\link{glottocheck}()}
}
\concept{<glottocheck>}
