% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkglottodata.R
\name{checkglottosubdata}
\alias{checkglottosubdata}
\title{Quality check of user-provided glottosubdata}
\usage{
checkglottosubdata(glottosubdata)
}
\arguments{
\item{glottosubdata}{}
}
\value{

}
\description{
Quality check of user-provided glottosubdata
}
\examples{
glottosubdata <- get_glottodata(meta = FALSE, dummy = "glottosubdata")
checkglottosubdata(glottosubdata[[1]])
lapply(glottosubdata, checkglottosubdata)
}
