% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkglottodata.R
\name{checkmetadata}
\alias{checkmetadata}
\title{Check metadata of a glottodataset}
\usage{
checkmetadata(glottodata)
}
\arguments{
\item{glottodata}{}
}
\value{

}
\description{
Check metadata of a glottodataset
}
\examples{
glottodata <- get_glottodata()
}
