% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createglottodata.R
\name{createglottodata}
\alias{createglottodata}
\title{Create empty glottodata for specified glottocodes and variables.}
\usage{
createglottodata(glottocodes, variables, filename = NULL, meta = TRUE, ...)
}
\arguments{
\item{glottocodes}{Character vector of glottocodes}

\item{variables}{Either a vector with variable names, or a single number indicating the total number of variable columns to be generated}

\item{filename}{Optional name of excel file where to store glottodata}

\item{...}{Other parameters passed to create_readmesheet(maintainer, email, citation, url)}
}
\value{

}
\description{
Output is a list with one data.frame for glottodata (and a number of metasheets if meta = TRUE). Output can  also be saved as an excel file.
}
\examples{
createglottodata(glottocodes = c("yucu1253", "tani1257"), variables = 3, filename = "glottodata.xlsx")
createglottodata(glottocodes = c("yucu1253", "tani1257"), variables = 3, filename = "glottodata_simple.xlsx", meta = FALSE)
}
\seealso{
Other <getdata>: 
\code{\link{createglottosubdata}()}
}
\concept{<getdata>}
