% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createglottodata.R
\name{createglottodata}
\alias{createglottodata}
\title{Create empty glottodata for specified glottocodes and variables.}
\usage{
createglottodata(glottocodes, variables, filename, ...)
}
\arguments{
\item{glottocodes}{Character vector of glottocodes}

\item{variables}{Either a vector with variable names, or a single number indicating the total number of variable columns to be generated}

\item{...}{Other parameters passed to create_readmesheet(maintainer, email, citation, url)}
}
\value{

}
\description{
Output is saved as an excel file with the following sheets: glottodata, structure, metadata, README
}
\examples{
createglottodata(glottocodes = c("yucu1253", "abcd1234"), variables = 3, filename = "glottodata.xlsx")
}
