% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createglottodata.R
\name{createglottosubdata}
\alias{createglottosubdata}
\title{Create empty glottosubdata for specified glottocodes, groups and variables.}
\usage{
createglottosubdata(
  glottocodes,
  variables,
  filename = NULL,
  groups,
  n = NULL,
  meta = TRUE,
  ...
)
}
\arguments{
\item{glottocodes}{Character vector of glottocodes}

\item{variables}{Either a vector with variable names, or a single number indicating the total number of variable columns to be generated}

\item{groups}{Character vector of group names}

\item{n}{Number of records to be assigned to each group}

\item{...}{Other parameters passed to create_readmesheet(maintainer, email, citation, url)}
}
\value{

}
\description{
glottosubcodes are created in the following form: glottocode_group_record.
For example: abcd1234_aaa_0001, abcd1234_aaa_0002, abcd1234_bbb_0001, abcd1234_bbb_0002
}
\details{
Output is saved as an excel file with the following sheets: glottodata, structure, metadata, readme
}
\examples{
createglottosubdata(glottocodes = c("yucu1253", "tani1257"), variables = 3, groups = c("a", "b"), n = 5, filename = "glottosubdata.xlsx")
}
