% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dplace}
\alias{dplace}
\title{This is an internally stored version of D-PLACE data. Use glottoget("dplace", download = TRUE) to download the latest version}
\format{
A data frame based on D-PLACE, the Database of Places, Language, Culture and Environment
}
\source{
\url{https://d-place.org/
}
}
\usage{
dplace
}
\description{
This is an internally stored version of D-PLACE data. Use glottoget("dplace", download = TRUE) to download the latest version
}
\keyword{internal}
\note{version: v2.2.1


This dataset contains the data from D-PLACE, the Database of Places, Language, Culture and Environment, serialized as CLDF StructureDataset.

##------ Thu Nov 13 17:57:29 2025 ------##}
