% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodataextract.R
\name{extractgeodata}
\alias{extractgeodata}
\alias{addgeodata}
\title{Extract and add environmental covariates}
\usage{
extractgeodata(glottodata, geodata, radius = NULL, fun = NULL, add = TRUE)
}
\arguments{
\item{glottodata}{Spatial vector with glottopoints or glottopols}

\item{geodata}{Spatial data in raster or vector format, or a path to such a file.}

\item{radius}{Radius (km) from which raster values should be extracted.}

\item{fun}{One of mean, median, min, max, sd, modal. Alternatively, a user-defined function can be specified, as in fun = function(x){mean(x,na.rm=T)}}

\item{add}{By default, extracted values are added to glottodata.}
}
\value{

}
\description{
Extract and add environmental covariates
}
\examples{
path <- "D:/Global/Societal/travel time/travel_time_to_cities_1.tif"
geodata <- get_geodata(path)

# extract from points
glottopoints <- glottofilter(continent = "South America")
glottopoints <- glottopoints[1:3, ]
extractgeodata(glottodata = glottopoints, geodata = geodata, radius = 10, fun = mean)

# extract from polygons
glottopoints <- glottofilter(country = "Netherlands")
glottopols <- points2pols(glottopoints, interpolation = "voronoi", country = "Netherlands")
extractgeodata(glottodata = glottopols, geodata = geodata,fun = mean)
}
