% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoget.R
\name{geoget}
\alias{geoget}
\title{Get geographic data}
\usage{
geoget(
  geodata,
  crs = NULL,
  attributes = FALSE,
  continent = NULL,
  country = NULL,
  region = NULL,
  union = FALSE,
  plot = FALSE,
  antarctica = FALSE
)
}
\arguments{
\item{geodata}{Either a filepath to locally stored geodata, or the name of geodata to download ("elevation", "basemap", or "climate").}

\item{crs}{Coordinate Reference System to transform to (default is Eckert IV).}

\item{attributes}{Whether to keep polygon attributes, or return only geometries}

\item{continent}{Optional filter by continent}

\item{country}{Optional filter by country}

\item{region}{Optional filter by region}

\item{union}{Should the polygons be merged into a single polygon with resolved boundaries?}

\item{plot}{Should the resulting polygon(s) be plotted}

\item{antarctica}{Should Antarctica be kept?}
}
\value{

}
\description{
Load spatial data either from local path (supports both raster and vector
formats), or download from remote server.
}
\examples{
From local path (not run):
geoget(geodata = "D:/data/vector.shp")
geoget(geodata = "D:/data/raster.tif")

Download:
geoget(geodata = "elevation", country = "Netherlands")
geoget(geodata = "climate")
}
\seealso{
Other <geodata>: 
\code{\link{geoget_basemap}()},
\code{\link{geoget_path}()},
\code{\link{geoget_remote}()}
}
\concept{<geodata>}
