% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodata.R
\name{get_geodata}
\alias{get_geodata}
\title{Get spatial data}
\usage{
get_geodata(path = NULL, download = NULL)
}
\arguments{
\item{path}{Path to geodata}

\item{download}{Name of geodata to download (wrapper around raster::getData). Currently, the following are supported: "climate"}
}
\value{

}
\description{
Load spatial data either from local path (supports both raster and vector formats), or download from remote server.
}
