% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodata.R
\name{get_geodata}
\alias{get_geodata}
\title{Open geodata (raster or vector format)}
\usage{
get_geodata(path, name = NULL)
}
\arguments{
\item{path}{Path to raster or vector file}

\item{name}{Optional}
}
\value{

}
\description{
Open geodata (raster or vector format)
}
