% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geodata.R
\name{get_geodata_path}
\alias{get_geodata_path}
\title{Get spatial data from local path}
\usage{
get_geodata_path(path)
}
\arguments{
\item{path}{Path to raster (RasterLayer or RasterStack) or vector data}
}
\value{

}
\description{
Open spatial data from a local path (supports both raster and vector formats).
}
