% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottodata_glottolog.R
\name{get_glottolog}
\alias{get_glottolog}
\title{Download glottolog data and convert it into a glot or geoglot object}
\usage{
get_glottolog(data = "glottolog_spatial")
}
\arguments{
\item{data}{One of \code{"glottolog"}, \code{"glottolog_spatial"}, or
\code{"glottolog_cldf"} or \code{"glottolog_cldf_spatial"}. By default, the downloaded glottolog data is
converted into a geoglot object (languages without coordinates are removed).}
}
\value{
Either a glot or geoglot object.
}
\description{
Download glottolog data and convert it into a glot or geoglot object
}
\examples{
glottodata <- get_glottolog()
}
