% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottodata_user.R
\name{get_sheetdata}
\alias{get_sheetdata}
\title{Import sheets from an excel file}
\usage{
get_sheetdata(path, sheets = NULL, simplify = TRUE)
}
\arguments{
\item{path}{Path to excel file (.xls or .xlsx)}

\item{sheets}{Character vector of sheetnames to import. Leave empty in case you want to import all sheets.}

\item{simplify}{If TRUE and only one sheet is selected, the data will be returned without placing it inside a list (which would be a list of length one).}
}
\value{
Named list with one table (tibble) per glot.
}
\description{
Import sheets from an excel file (.xls or .xlsx).
Use the function \code{"get_sheetnames"} to show all sheetnames in an excel file (without importing them).
}
\details{
This is a wrapper around excel_sheets.
}
\examples{
# 1. There is one sheet in which each row represents a glot (anything with a glottocode). The first column contains the glottocodes.
get_sheetdata(path = "glottodata.xlsx")
get_sheetdata(path = "glottodata.xlsx", sheets = 1)
get_sheetdata(path = "glottodata.xlsx", sheets = "glottodata")

path <- "C:/Users/sjnor/surfdrive/Shared/SAPPHIRE/Output/Presentations/Isolates_WS-Athens/data/dbase_isolates_V2.xlsx"
data <- get_sheetdata(path = path, sheets = "isolates")

# 2. There are several sheets, each of which represents a glot. The name of the sheets are glottocodes.  The first column in each sheet contains glottosubcodes (abcd1234_a1)
get_sheetdata(path = "glottodata.xlsx", sheets = c("abcd1234", "abcd1235", "abcd1236") )
get_sheetdata(path = "glottodata.xlsx", sheets = c(1:16) )

path <- "C:/Users/sjnor/surfdrive/Shared/Tanimuka_Yucuna/Databases/Voor_Sietze/NounPhrase_V6.xlsx"
get_sheetnames(path)
data <- get_sheetdata(path = path, sheets = c(1:16) )

# 3. There are several sheets, each of which represents a subglot. The name of the sheets are glottosubcodes. The first column in each sheet contains glottosubsubcodes (abcd1234_a1_a1)
get_sheetdata(path = "glottodata.xlsx", sheets = c("abcd1234_form", "abcd1234_meaning", "abcd1235_form", "abcdf1235_meaning") )

path <- "C:/Users/sjnor/surfdrive/Shared/Tanimuka_Yucuna/Databases/Voor_Sietze/TAME_V6.xlsx"
get_sheetnames(path)
}
