% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottodata_user.R
\name{get_sheetnames}
\alias{get_sheetnames}
\title{List names of all sheets in an excel file}
\usage{
get_sheetnames(path)
}
\arguments{
\item{path}{Path to excel file}
}
\value{
Vector of names
}
\description{
Reads an excel file (.xls or .xlsx).
Use the function \code{"get_sheetdata"}  to import data.
}
\details{
This is a wrapper around excel_sheets.
}
\examples{
get_sheetnames(path = "glottodata.xlsx")
get_sheetnames(path = "glottodata.xls")
}
