% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottosearch.R
\name{glot_exists}
\alias{glot_exists}
\title{Check whether a glot or vector of glots exist in glottolog}
\usage{
glot_exists(find, columns)
}
\arguments{
\item{find}{Glottocode, name of language, family, etc.}

\item{columns}{In which column should be searched}
}
\value{
Logical: TRUE/FALSE
}
\description{
Check whether a glot or vector of glots exist in glottolog
}
\examples{
#' glot_exists(find = c("yucu1253", "abcd1234"), columns = "glottocode")
}
