% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{glot_exists}
\alias{glot_exists}
\title{Check whether a glottocode, language, family, or dialect exists in glottolog. This is a wrapper around glotto_search (only exact matches are returned)}
\usage{
glot_exists(find, columns = NULL)
}
\value{

}
\description{
Check whether a glottocode, language, family, or dialect exists in glottolog. This is a wrapper around glotto_search (only exact matches are returned)
}
