% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottosearch.R
\name{glot_exists}
\alias{glot_exists}
\title{Check whether a glot or vector of glots exist in glottolog}
\usage{
glot_exists(find, columns)
}
\arguments{
\item{find}{Glottocode, name of language, family, etc.}

\item{columns}{In which column should be searched}
}
\value{
Logical: TRUE/FALSE
}
\description{
This function checks for exact matches, and returns TRUE/FALSE. In case you
want a more flexible search, you can use
\code{\link[=glottosearch]{glottosearch()}}, in case you want to check
whether glottocodes exist/are valid, use
\code{\link[=glottocodes_exist]{glottocodes_exist()}}
}
\examples{
glot_exists(find = c("yucu1253", "abcd1234"), columns = "glottocode")
}
\seealso{
Other <glottocheck><glottosearch>: 
\code{\link{glottocode_exists}()},
\code{\link{glottocode_online}()},
\code{\link{glottosubcode_valid}()}
}
\concept{<glottocheck><glottosearch>}
