% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glottobase}
\alias{glottobase}
\title{Load glottobase data}
\format{
A data frame with XX rows and XX variables:
\describe{
  \item{var1}{explain}
  \item{var2}{explain}
  ...
}
}
\source{
\url{http://www.glottolog.org/
}
}
\usage{
glottobase
}
\description{
This is a boosted/enriched version of glottolog data. It is used as reference dataset in several functions. The glottocodes are used as unique IDs to match different data sources.
}
\keyword{datasets}
