% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glottobase}
\alias{glottobase}
\title{glottobase data}
\format{
A data frame with XX rows and XX variables:
\describe{
  \item{glottocode}{description}
  \item{family_id}{description}
  \item{parent_id}{description}
  \item{name}{description}
  \item{isocode}{description}
  \item{child_dialect_count}{description}
  \item{country_ids}{description}
  \item{family_name}{description}
  \item{isolate}{description}
  \item{family_size}{description}
  \item{family_size_rank}{description}
  \item{country}{description}
  \item{continent}{description}
  \item{region}{description}
  \item{geometry}{description}
  ...
}
}
\source{
\url{https://glottolog.org/
}
}
\usage{
glottobase
}
\description{
This is an internally stored and spatially enhanced version of glottolog data. It is included in the package as a fall-back option in case there is no internet connection. It is used as reference dataset in several functions. The glottocodes are used as unique IDs to match different data sources.
}
\keyword{internal}
