% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottobooster.R
\name{glottobooster}
\alias{glottobooster}
\title{Glottologbooster: enhance glottolog data}
\usage{
glottobooster(
  glottologdata = NULL,
  space = TRUE,
  addfamname = TRUE,
  addisolates = TRUE,
  L1only = TRUE,
  addfamsize = TRUE,
  addfamsizerank = TRUE
)
}
\arguments{
\item{glottologdata}{data from \href{https://glottolog.org/}{glottolog}, can be downloaded with \code{\link{glottoget_glottolog()}}}

\item{space}{Return spatial object?}

\item{addfamname}{Add column with familiy names?}

\item{addisolates}{Add column to identify isolates?}

\item{L1only}{Keep only L1 languages (remove bookkeeping, unclassifiable, sign languges, etc.). See glottobooster_rm for more flexibility.}

\item{addfamsize}{Add column with family size?}

\item{addfamsizerank}{Add column with family size rank?}
}
\value{

}
\description{
Glottologbooster: enhance glottolog data
}
\examples{
glottologdata <- glottoget("glottolog")
glottologdata <- glottobooster(glottologdata)
}
\keyword{internal}
