% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocheck.R
\name{glottocheck_hasmeta}
\alias{glottocheck_hasmeta}
\alias{checkmetadata_hasstructure}
\title{Check whether glottodata contains metadata}
\usage{
glottocheck_hasmeta(glottodata)
}
\arguments{
\item{glottodata}{}
}
\value{

}
\description{
In fact, this function only checks whether glottodata contains a structure table, because the structure table is the only table that is required by some glottospace functions. All other tables are for humans, not computers;-).
}
\seealso{
Other <glottocheck>: 
\code{\link{checkglottodata}()},
\code{\link{checkglottosubdata}()},
\code{\link{glottocheck}()}
}
\concept{<glottocheck>}
