% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocheck.R
\name{glottocheck_isglottodata}
\alias{glottocheck_isglottodata}
\title{Guess whether a list of glottodata tables is glottodata (and not glottosubdata)}
\usage{
glottocheck_isglottodata(glottodata)
}
\arguments{
\item{glottodata}{User-provided glottodata}
}
\value{

}
\description{
Guess whether a list of glottodata tables is glottodata (and not glottosubdata)
}
\examples{
glottocheck_isglottodata(glottodata)
}
