% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocheck.R
\name{glottocheck_isglottosubdata}
\alias{glottocheck_isglottosubdata}
\title{Guess whether a dataset is glottosubdata}
\usage{
glottocheck_isglottosubdata(glottodata)
}
\arguments{
\item{glottodata}{User-provided glottodata}
}
\value{

}
\description{
Guess whether a dataset is glottosubdata
}
