% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocheck.R
\name{glottocheck_isglottosubdata}
\alias{glottocheck_isglottosubdata}
\title{Guess whether an object is glottosubdata (and not glottodata)}
\usage{
glottocheck_isglottosubdata(glottodata)
}
\arguments{
\item{glottodata}{User-provided glottodata}
}
\description{
Guess whether an object is glottosubdata (and not glottodata)
}
\examples{
glottocheck_isglottosubdata(glottodata)
}
\keyword{internal}
