% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocheck.R
\name{glottocheck_subdata}
\alias{glottocheck_subdata}
\title{Quality check of user-provided glottosubdata}
\usage{
glottocheck_subdata(glottosubdata, diagnostic = TRUE)
}
\arguments{
\item{glottosubdata}{User-provided glottosubdata}

\item{diagnostic}{If TRUE (default) a data viewer will be opened to show the levels of each variable (including NAs), and a data coverage plot will be shown.}
}
\value{

}
\description{
Quality check of user-provided glottosubdata
}
\examples{
glottosubdata <- glottoget_glottodata(meta = FALSE, demo = "glottosubdata")
glottosubdata <- join_glottodata(glottosubdata)
glottocheck_subdata(glottosubdata)

Better to join_glottodata first, instead of the following approach (because checks only within each language for duplicates etc.)
lapply(glottosubdata, glottocheck_subdata)
}
\seealso{
Other <glottocheck>: 
\code{\link{glottocheck_data}()},
\code{\link{glottocheck_hasmeta}()},
\code{\link{glottocheck}()}
}
\concept{<glottocheck>}
