% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocheck.R
\name{glottocheck_varlevels}
\alias{glottocheck_varlevels}
\title{Show levels of all variables.}
\usage{
glottocheck_varlevels(data, diagnostic = TRUE)
}
\arguments{
\item{data}{Glottodata}

\item{diagnostic}{Whether diagnostic messages should be shown}
}
\value{
Opens a data viewer showing the levels of each variable (including NA)
}
\description{
This function shows the levels of all variables and is mainly intended to provide users with a quick overview of their data.
}
\examples{
glottodata <- glottoget("demodata", meta = FALSE)
glottocheck_varlevels(data = glottodata, diagnostic = FALSE)
}
\keyword{internal}
