% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocode.R
\name{glottocode}
\alias{glottocode}
\title{Get detailed info for a specific glottocode}
\usage{
glottocode(glottocode)
}
\arguments{
\item{glottocode}{A single glottocode}
}
\value{
Diagnostic messages indicating whether a glottocode exists,
a plot showing the location of the glottocode on the Earth,
a pop-up browser window to the Glottolog webpage.
}
\description{
This function does three things for a glottocode:
- Plot its location on the Earth
- Open the associated webpage on www.glottolog.org
- Check whether it exists in glottolog
}
\examples{
\donttest{
glottocode("yucu1253")
}
}
