% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottosearch.R
\name{glottocode_exists}
\alias{glottocode_exists}
\alias{glottocodes_exist}
\title{Check whether glottocodes exist in glottolog}
\usage{
glottocode_exists(glottocode)
}
\arguments{
\item{glottocode}{A glottocode or character vector of glottocodes}
}
\value{
A logical vector
}
\description{
Check whether glottocodes exist in glottolog
}
\examples{
glottocode_exists(c("yucu1253"))
glottocode_exists(c("yucu1253", "abcd1234"))
}
\seealso{
Other <glottocheck><glottosearch>: 
\code{\link{glot_exists}()},
\code{\link{glottocode_online}()},
\code{\link{glottosubcode_valid}()}
}
\concept{<glottocheck><glottosearch>}
