% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocode.R
\name{glottocode_exists}
\alias{glottocode_exists}
\title{Check whether a set of glottocodes exist in glottolog}
\usage{
glottocode_exists(glottocode)
}
\arguments{
\item{glottocode}{A glottocode or character vector of glottocodes}
}
\value{
A logical vector
}
\description{
Checks whether a set of glottocodes exist in glottolog (checked at the level of L1 languages)
}
\examples{
\donttest{
glottocode_exists(c("yucu1253"))
glottocode_exists(c("yucu1253", "abcd1234"))
}
}
\concept{<glottocheck><glottosearch>}
