% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocode.R
\name{glottocode_online}
\alias{glottocode_online}
\title{Open url in web browser for glottocode}
\usage{
glottocode_online(glottocode)
}
\arguments{
\item{glottocode}{}
}
\description{
Open url in web browser for glottocode
}
\examples{
glottocode_online("yucu1253")
}
\seealso{
Other <glottocheck><glottosearch>: 
\code{\link{glottocode_exists}()},
\code{\link{glottosearch_exist}()},
\code{\link{glottosubcode_valid}()}
}
\concept{<glottocheck><glottosearch>}
\keyword{internal}
