% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoconvert.R
\name{glottoconvert}
\alias{glottoconvert}
\title{glottoconvert}
\usage{
glottoconvert(
  data,
  glottocode = NULL,
  var,
  table = NULL,
  ref = NULL,
  page = NULL,
  remark = NULL,
  contributor = NULL
)
}
\arguments{
\item{data}{Dataset to be transformed to glottodata}

\item{glottocode}{column name or column id with glottocodes}

\item{var}{Character string that distinguishes those columns which contain variable names.}

\item{table}{In case dataset consists of multiple tables, indicate which table contains the data that should be converted.}

\item{ref}{Character string that distinguishes those columns which contain references.}

\item{page}{Character string that distinguishes those columns which contain page numbers.}

\item{remark}{Character string that distinguishes those columns which contain remarks.}

\item{contributor}{Character string that distinguishes those columns which contain contributors.}
}
\value{

}
\description{
glottoconvert
}
\examples{
\dontrun{
data <- glottoget("userdata.xlsx")
glottodata <- glottoconvert(data = data, var = "var_", ref = "ref_", page = "pag_", remark = "com_", contributor = "Coder")
glottocheck(glottodata)
glottosave(glottodata)
}
}
