% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoconvert.R
\name{glottoconvert}
\alias{glottoconvert}
\title{glottoconvert}
\usage{
glottoconvert(
  data,
  glottocode = NULL,
  table = NULL,
  var = NULL,
  ref = NULL,
  page = NULL,
  position = NULL
)
}
\arguments{
\item{data}{Dataset to be transformed to glottodata}

\item{glottocode}{column name or column id with glottocodes}

\item{table}{In case dataset consists of multiple tables, indicate which table contains the data that should be converted.}

\item{var}{Character string that distinguishes those columns which contain variable names.}

\item{ref}{Character string that distinguishes those columns which contain references.}

\item{page}{Character string that distinguishes those columns which contain page numbers.}

\item{position}{Optional position of distinguishing character string, either "start" or "end"}
}
\value{

}
\description{
glottoconvert
}
\examples{
# Load data from path:
# data <- glottoget("C:/Users/sjnor/surfdrive/Projecten en schrijfsels/Papers in progress/Yucuna-Tanimuca/data/testdata linguistics/2022_01_16_basic_data.xlsx")
}
