% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoconvert.R
\name{glottoconvert_subdata}
\alias{glottoconvert_subdata}
\title{Convert subdata into glottosubdata}
\usage{
glottoconvert_subdata(subdata, glottocodes = NULL)
}
\arguments{
\item{subdata}{Any dataset that should be converted into glottosubdata. This will generally an excel file loaded with glottoget() of which the sheetnames are glottocodes.}

\item{glottocodes}{Optional character vector of glottocodes. If no glottocodes are supplied, glottospace will search for them in the sample table.}
}
\description{
Convert subdata into glottosubdata
}
