% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocreate.R
\name{glottocreate_sampletable}
\alias{glottocreate_sampletable}
\title{create sample table for glottodata}
\usage{
glottocreate_sampletable(glottocodes = NULL)
}
\arguments{
\item{glottocodes}{Character vector of glottocodes}
}
\value{
A sample table
}
\description{
create sample table for glottodata
}
\examples{
glottocreate_sampletable(glottocodes = c("yucu1253", "tani1257"))
}
\keyword{internal}
