% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottocreate.R
\name{glottocreate_subdata}
\alias{glottocreate_subdata}
\title{Create empty glottosubdata for specified glottocodes, groups and variables.}
\usage{
glottocreate_subdata(
  glottocodes,
  variables,
  groups,
  filename = NULL,
  check = FALSE,
  levels = NULL,
  n = NULL,
  meta = TRUE,
  maintainer = NULL,
  email = NULL,
  citation = NULL,
  url = NULL
)
}
\arguments{
\item{glottocodes}{Character vector of glottocodes}

\item{variables}{Either a vector with variable names, or a single number indicating the total number of variable columns to be generated}

\item{groups}{Character vector of group names}

\item{filename}{Optional name of excel file where to store glottodata}

\item{check}{Should glottocodes be checked? Default is FALSE because takes much time to run.}

\item{levels}{Optional character vector with levels across all variables}

\item{n}{Number of records to be assigned to each group}

\item{meta}{Should metatables be created?}

\item{maintainer}{Name of the person/organization maintaining the data (optional)}

\item{email}{Email address of maintainer/contact person (optional)}

\item{citation}{How to cite the data (optional)}

\item{url}{Optional url linking to a webpage.}
}
\value{
A table or list of tables
}
\description{
glottosubcodes are created in the following form: glottocode_group_record.
For example: abcd1234_aaa_0001, abcd1234_aaa_0002, abcd1234_bbb_0001, abcd1234_bbb_0002
}
\details{
Output is a list with one data.frame per language (and a number of metatables if meta = TRUE). Output can  also be saved as an excel file.
}
\examples{
glottocreate_subdata(glottocodes = c("yucu1253", "tani1257"),
variables = 3, groups = c("a", "b"), n = 5, filename = "glottosubdata.xlsx")
}
\concept{<glottoget_path><glottocreate>}
\keyword{internal}
