% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottospace.R
\name{glottodata_addcountries}
\alias{glottodata_addcountries}
\alias{glottodata_addcontinents}
\alias{glottodata_addregions}
\title{Add names of countries, continents, and regions to glottodata}
\usage{
glottodata_addcountries(glottodata)
}
\arguments{
\item{glottodata}{User-provided glottodata}
}
\value{
A spatial (sf) object
}
\description{
Spatial intersection between glottodata and
\href{https://www.naturalearthdata.com/}{naturalearthdata} and extracts names
of countries, continents, and regions.
}
\examples{
glottodata <- glottoget_glottodata(meta = FALSE)
glottodata_addcountries(glottodata)
}
