% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottodist.R
\name{glottodist}
\alias{glottodist}
\title{Calculate distances between languages}
\usage{
glottodist(glottodata, structure = NULL)
}
\arguments{
\item{glottodata}{A glottodata table, or a list with a glottodata table and a structure table.}

\item{structure}{If glottodata is a table without metadata, a structure table should be provided. You can create it with glottocreate_structuretable() and add it with glottodata_addtable()}
}
\value{
object of class \code{dist}
}
\description{
Calculate distances between languages
}
\examples{
glottodata <- glottoget("demodata", meta = TRUE)
glottodist <- glottodist(glottodata = glottodata)
}
