% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottodist.R
\name{glottodist}
\alias{glottodist}
\title{Calculate distances between languages}
\usage{
glottodist(
  glottodata,
  types = NULL,
  levels = NULL,
  weights = NULL,
  structure = NULL
)
}
\arguments{
\item{types}{Character vector with the same length as the number of columns specifying the type of each column}

\item{levels}{Character vector with the same length as the number of columns specifying the levels of each column}

\item{weights}{Character vector with the same length as the number of columns specifying the weight of each column}

\item{structure}{Data frame specifying per column in the data (optional): types, levels weights. Columns should be named as follows: colnames, type, weight.}

\item{data}{Data frame. Rows can either be glots, or subglots (lower-level aspects of a language, e.g. constructions).
Columns contain the variables based on which distances are calculated.}
}
\value{
object of class \code{dist}
}
\description{
Calculate distances between languages
}
\examples{
glottodata <- glottoget(meta = TRUE)
glottodist <- glottodist(glottodata = glottodata)
}
