% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottodist.R
\name{glottodist}
\alias{glottodist}
\title{Calculate distances between languages}
\usage{
glottodist(glottodata, metric = "gower")
}
\arguments{
\item{glottodata}{glottodata or glottosubdata, either with or without structure table.}
}
\value{
object of class \code{dist}
}
\description{
Calculate distances between languages
}
\examples{
glottodata <- glottoget("demodata", meta = TRUE)
glottodist <- glottodist(glottodata = glottodata)

glottosubdata <- glottoget("demosubdata", meta = TRUE)
glottodist <- glottodist(glottodata = glottosubdata)
# glottoplot(glottodist)
}
