% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottodist.R
\name{glottodist_subdata}
\alias{glottodist_subdata}
\title{Calculate indexing distances between languages}
\usage{
glottodist_subdata(
  glottosubdata,
  metric = NULL,
  index_type = NULL,
  avg_idx = NULL,
  fixed_idx = NULL
)
}
\arguments{
\item{glottosubdata}{an glottosubdata object}

\item{metric}{either "gower" or "anderberg"}

\item{index_type}{either "mc" or "relative" or "fmi"}

\item{avg_idx}{the feature indices over which the average of distances is computed, it must be given when index_type is either "relative" or "fmi".}

\item{fixed_idx}{the feature indices over which the distance of two constructions is computed, it must be given when index_type is either "relative" or "fmi".}
}
\value{
object of class \code{dist}
}
\description{
Calculate indexing distances between languages
}
\section{Details}{

The function ``glottodist_subdata'' returns a ``dist'' object,
the input is a glottosubdata object,
it computes the indexing between languages,
We refer to the observations of each language as constructions.
The distance \eqn{d(A_i, B_j)} between two construction \eqn{A_i} in a language \eqn{A} and \eqn{B_j} in a language \eqn{B}
is determined by the argument ``metric'',
whose value is either ``gower'' or ``anderberg''.
When ``index'' is ``mc'',
it returns the ``matching constructions'':

\eqn{MC(A, B) := \frac{1}{2|A|}\sum\limits_{A_i\in A}\min\limits_{B_j\in B}d(A_i, B_j) +
\frac{1}{2|B|}\sum\limits_{B_i\in B}\min\limits_{A_j\in A}d(A_j, B_i)}.
When ``metric'' is ``relative'',
it returns the ``relative indexing'':

\eqn{RI(A, B) = \frac{1}{|M|}\sum\limits_{s\in M}\textrm{AVG}_{A_i(s) = 1 \textrm{ and } B_j(s) = 1}d(A_i^F, B_j^F)},
here \eqn{M} is a subset of variables given by the argument ``avg_idx'' and \eqn{F} is a subset of variables given by the argument ``fixed_idx'',
and \eqn{A_i^F}, \eqn{B_j^F} are given by the constructions \eqn{A_i}, \eqn{B_j} restricted to ``fixed_idx'' \eqn{F}.
When ``metric'' is ``fmi'',
it returns the ``form-meaning indexing'':

\eqn{FMI(A, B) = \frac{1}{|M||F|} \sum\limits_{s\in M, p\in F} SIM(\{(A_i^M(s)=1 \textrm{ and }A_i^F(p)=1)\},
\{B_j^M(s) = 1 \textrm{ and }B_j^F(p) = 1\})},
here \eqn{SIM(X, Y)=\min(|X|/|Y|, |Y|/|X|)}, if both \eqn{X} and \eqn{Y} are empty,
\eqn{SIM(X, Y)=1}.
}

\examples{
glottosubdata_cnstn <- glottoget(glottodata = "demosubdata_cnstn")
glottodist_subdata(glottosubdata = glottosubdata_cnstn, metric = "gower", index = "mc")
glottodist_subdata(glottosubdata = glottosubdata_cnstn, metric = "gower", index = "relative",
                   avg_idx = 1:4, fixed_idx = 5:7)
glottodist_subdata(glottosubdata = glottosubdata_cnstn, index = "fmi",
                   avg_idx = 1:4, fixed_idx = 5:7)

}
