% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottodist.R
\name{glottodist_subdata}
\alias{glottodist_subdata}
\title{Calculate indexing distances between languages}
\usage{
glottodist_subdata(
  glottosubdata,
  metric = NULL,
  index_type = NULL,
  avg_idx = NULL,
  fixed_idx = NULL
)
}
\arguments{
\item{glottosubdata}{an glottosubdata object}

\item{metric}{either "gower" or "anderberg"}

\item{index_type}{either "mc" or "relative" or "fmi"}

\item{avg_idx}{the feature indices over which the average of distances is computed, it must be given when index_type is either "relative" or "fmi".}

\item{fixed_idx}{the feature indices over which the distance of two constructions is computed, it must be given when index_type is either "relative" or "fmi".}
}
\value{
object of class \code{dist}
}
\description{
Calculate indexing distances between languages
}
