% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottodrop.R
\name{glottodrop}
\alias{glottodrop}
\title{Drop glottodata tables and properties}
\usage{
glottodrop(
  glottodata,
  dropmeta = TRUE,
  dropspatial = TRUE,
  dropsub = TRUE,
  droplist = TRUE,
  tbdf = TRUE
)
}
\arguments{
\item{glottodata}{glottodata or glottosubdata.}

\item{dropmeta}{By default all metadata is dropped.}

\item{dropspatial}{By default spatial properties are dropped.}

\item{dropsub}{By default, glottosubdata tables are merged into a single glottodata table.}

\item{droplist}{If glottodata is a list, should the glottodata table be returned?}

\item{tbdf}{If glottodata is a tibble, should it be converted to a data.frame?}
}
\value{

}
\description{
Drop glottodata tables and properties
}
\examples{
glottodrop(glottodata)
}
