% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottofilter.R
\name{glottofilter}
\alias{glottofilter}
\alias{glotto_select}
\alias{glotto_subset}
\title{Filter glottolog data}
\usage{
glottofilter(
  glottodata = NULL,
  isocodes = NULL,
  glottocode = NULL,
  family_name = NULL,
  family_id = NULL,
  continent = NULL,
  country = NULL,
  expression = NULL
)
}
\arguments{
\item{glottodata}{glot or geoglot object obtained with
\code{\link{glotto_get}} or \code{\link{as_glot}}  or
\code{\link{as_geoglot}}.}

\item{isocodes}{A character vector of isocodes}

\item{glottocode}{A character vector of glottocodes}

\item{family_name}{A character vector of language families}

\item{family_id}{A character vector of language family IDs}

\item{continent}{A character vector of continents}

\item{country}{A character vector of countries}

\item{expression}{A regular expression}
}
\value{
A smaller glot object or geoglot object containing only the selected features.
}
\description{
Filter glottolog data
}
\examples{
# Get data
glottoraw <- gs_data(name = "glottolog")
glottodata <- gs_upgrade(data = glottoraw)

# Filter glottodata
points <- glottofilter(glottodata = glottodata, isocodes = colnames(distmat))
points <- glottofilter(glottodata = glottodata, glottocode = "wari1268")
points <- glottofilter(glottodata = glottodata, family_name = 'Indo-European')
points <- glottofilter(glottodata = glottodata, continent = "South America")
points <- glottofilter(glottodata = glottodata, country = c("Colombia", "Venezuela"))
points <- glottofilter(glottodata = glottodata, expression = family_name \%in\% c("Arawakan", "Tucanoan"))
}
