% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottofilter.R
\name{glottofilter}
\alias{glottofilter}
\title{Filter glottodata by language, glottocode, etc.}
\usage{
glottofilter(
  glottodata = NULL,
  isocodes = NULL,
  glottocode = NULL,
  family_name = NULL,
  family_id = NULL,
  continent = NULL,
  country = NULL,
  region = NULL,
  expression = NULL
)
}
\arguments{
\item{glottodata}{glot or geoglot object obtained with
\code{\link{glottoget_glottolog}} or \code{\link{as_glot}}  or
\code{\link{as_geoglot}}.}

\item{isocodes}{A character vector of iso639P3codes}

\item{glottocode}{A character vector of glottocodes}

\item{family_name}{A character vector of language families}

\item{family_id}{A character vector of language family IDs}

\item{continent}{A character vector of continents}

\item{country}{A character vector of countries}

\item{expression}{A regular expression}
}
\value{
A subset of the original glottodata table  containing only filtered languages.
}
\description{
By default, the glottolog data will be used to filter from. But in case the user provides glottodata, this will be used.
}
\examples{
# Get data
glottoraw <- gs_data(name = "glottolog")
glottodata <- gs_upgrade(data = glottoraw)

# Filter glottodata
points <- glottofilter(glottodata = glottodata, isocodes = colnames(distmat))
points <- glottofilter(glottodata = glottodata, glottocode = "wari1268")
points <- glottofilter(glottodata = glottodata, family_name = 'Indo-European')
points <- glottofilter(glottodata = glottodata, continent = "South America")
points <- glottofilter(glottodata = glottodata, country = c("Colombia", "Venezuela"))
points <- glottofilter(glottodata = glottodata, expression = family_name \%in\% c("Arawakan", "Tucanoan"))
}
\concept{<glottofilter><glottosearch>}
