% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottofilter.R
\name{glottofilter}
\alias{glottofilter}
\title{Filter glottodata by language, glottocode, etc.}
\usage{
glottofilter(
  glottodata = NULL,
  isocodes = NULL,
  glottocode = NULL,
  name = NULL,
  family = NULL,
  family_id = NULL,
  continent = NULL,
  country = NULL,
  region = NULL,
  expression = NULL,
  colname = NULL,
  keep = NULL,
  drop = NULL,
  select = NULL
)
}
\arguments{
\item{glottodata}{A glottodata table}

\item{isocodes}{A character vector of iso639p3codes}

\item{glottocode}{A character vector of glottocodes}

\item{name}{A character vector of language names}

\item{family}{A character vector of language families}

\item{family_id}{A character vector of language family IDs}

\item{continent}{A character vector of continents}

\item{country}{A character vector of countries}

\item{region}{A character vector of regions}

\item{expression}{A regular expression}

\item{colname}{A column name}

\item{keep}{Character vector of things to keep (only if colname is provided)}

\item{drop}{Character vector of things to keep (only if colname is provided)}
}
\value{
A subset of the original glottodata table  containing only filtered languages.
}
\description{
By default, the glottolog data will be used to filter from. But in case the user provides glottodata, this will be used.
}
\examples{
# Get data
glottoraw <- gs_data(name = "glottolog")
glottodata <- gs_upgrade(data = glottoraw)

# Filter glottodata
points <- glottofilter(glottodata = glottodata, isocodes = colnames(distmat))
points <- glottofilter(glottodata = glottodata, glottocode = "wari1268")
points <- glottofilter(glottodata = glottodata, family = "Indo-European")
points <- glottofilter(glottodata = glottodata, continent = "South America")
points <- glottofilter(glottodata = glottodata, family = "Indo-European", continent = "South America")
points <- glottofilter(glottodata = glottodata, country = c("Colombia", "Venezuela"))
points <- glottofilter(glottodata = glottodata, expression = family \%in\% c("Arawakan", "Tucanoan"))
points <- glottofilter(glottodata = glottodata, colname = "family", drop = "Indo-European")
}
\concept{<glottofilter><glottosearch>}
