% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottofilter.R
\name{glottofiltermap}
\alias{glottofiltermap}
\title{Filter languages from a map}
\usage{
glottofiltermap(glottodata = NULL, mode = NULL, ...)
}
\arguments{
\item{glottodata}{Spatial glottodata object}

\item{mode}{Either "draw" or "click"}

\item{...}{Additional arguments to pass to glottofilter}
}
\value{
A set of languages selected from the original glottodata object
}
\description{
Select languages by drawing or clicking on a map
}
\examples{
\dontrun{
selected <- glottofiltermap()
glottomap(selected)

glottofiltermap(continent = "South America")
glottofiltermap(country = "Netherlands")
}
}
