% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget.R
\name{glottoget}
\alias{glottoget}
\title{Get glottodata from local path or online global databases}
\usage{
glottoget(glottodata = NULL, meta = FALSE)
}
\arguments{
\item{glottodata}{options are:
\itemize{
\item A filepath to locally stored glottodata with file extension (.xlsx .xls .gpkg
.shp). See also: options meta and simplify.
\item "glottobase" - Default option, an enhanced version of \href{https://glottolog.org/}{glottolog}. See
\link{glottologbooster} for details.
\item "glottolog" - This is the raw data downloaded from \href{https://glottolog.org/}{glottolog}
\item "glottospace" - A simple dataset with glottocodes and a geometry column. This
is a subset of all languages in \href{https://glottolog.org/}{glottolog} with
spatial coordinates.
\item "demodata" - Built-in artificial glottodata (included for demonstration and testing)
\item "demosubdata" - Built-in artificial glottosubdata (included for demonstration and testing)
}}

\item{meta}{In case 'glottodata' is a path to locally stored data (or demodata/demosubdata): by default, meta sheets are not loaded. Use meta=TRUE if you want to include them.}

\item{simple}{In case 'glottodata' is a path to locally stored data  (or demodata/demosubdata): by default, if only one sheet is loaded, the data will be returned as a data.frame (instead of placing the data inside a list of length 1)}
}
\value{

}
\description{
Load locally stored glottodata, download databases from online sources, or load built-in demo data
}
\examples{
glottoget("glottolog")
}
\seealso{
Other <glottodata>: 
\code{\link{glottoget_path}()},
\code{\link{glottoget_remote}()},
\code{\link{glottosave}()}
}
\concept{<glottodata>}
