% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget.R
\name{glottoget}
\alias{glottoget}
\title{Get glottodata from local path or online global databases}
\usage{
glottoget(
  glottodata = NULL,
  meta = FALSE,
  download = FALSE,
  dirpath = NULL,
  url = NULL,
  seed = NULL
)
}
\arguments{
\item{glottodata}{options are:
\itemize{
\item A filepath to locally stored glottodata or glottosubdata with file extension (.xlsx .xls .gpkg
.shp). See also: options meta and simplify.
\item "glottobase" - Default option, an spatially enhanced version of \href{https://glottolog.org/}{glottolog}. See
\link{glottobooster} for details. If glottodata = NULL, "glottobase" will be loaded.
\item "wals" - A spatially enhanced version of \href{https://wals.info/}{WALS}.
\item "dplace" - A spatially enhanced version of \href{https://d-place.org/}{D-PLACE}.
\item "glottolog" - A restructured (non-spatial) version of \href{https://glottolog.org/}{glottolog}.
\item "glottospace" - A simple dataset with glottocodes and a geometry column.
\item "grambank" - A restructured (non-spatial) version of \href{https://grambank.clld.org/}{Grambank}.
\item "grambankspace" - A spatially enhanced version of Grambank.
\item "phoible_raw", "phoiblespace_raw", "phoible", "phoiblespace" - Variants of \href{https://phoible.org/}{PHOIBLE} (raw vs sampled, spatial vs non-spatial).
\item "phoible_raw_param_sf", "phoible_param_sf" - Returns sf object with parameter-wise spatial distributions for PHOIBLE.
\item "worldatlas" - Spatial data derived from Asher & Moseley (2007) via Zenodo \href{https://doi.org/10.5281/zenodo.15287258}{DOI}.
\item "demodata", "demosubdata", "demosubdata_cnstn" - Built-in artificial glottodata.
}}

\item{meta}{If TRUE, includes meta sheets when loading demodata/demosubdata.}

\item{download}{If TRUE, attempts to download latest data from Zenodo.}

\item{dirpath}{Optional path for storing or reading downloaded files.}

\item{url}{Zenodo URL (e.g., "https://zenodo.org/api/records/3260727").}

\item{seed}{Used for sampling in phoible datasets.}
}
\value{
A glottodata or glottosubdata object (a data.frame or list)
}
\description{
Load locally stored glottodata, download databases from online sources, or load built-in demo data
}
\examples{
\donttest{
glottoget("glottolog")
glottoget("worldatlas", download = TRUE)
}
}
\seealso{
Other <glottodata>: 
\code{\link{glottosave}()}
}
\concept{<glottodata>}
