% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget.R
\name{glottoget_glottolog}
\alias{glottoget_glottolog}
\title{Get glottolog data}
\usage{
glottoget_glottolog(days = NULL)
}
\value{

}
\description{
This function checks whether most recent version of glottolog is locally available. If local version is outdated, the newest version will be downloaded.
}
\examples{
glottoget_glottolog()
}
