% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget.R
\name{glottoget_path}
\alias{glottoget_path}
\title{Load glotto(sub)data from file}
\usage{
glottoget_path(filepath = NULL, simplify = TRUE)
}
\arguments{
\item{filepath}{Path to glottodata file with extension (.xlsx .xls .gpkg .shp). If no filepath is specified, an artificial demo dataset will be created.}

\item{simplify}{By default, if only one sheet is loaded, the data will be returned as a data.frame (instead of placing the data inside a list of length 1)}
}
\description{
Load glottodata/glottosubdata from a file
}
\examples{
\dontrun{
glottoget_path()
glottoget_path(filepath = "glottodata.xlsx")
glottoget_path(filepath = "glottodata.gpkg")
}
}
\seealso{
glottosave

Other <glottodata>: 
\code{\link{glottoget_remote}()},
\code{\link{glottoget}()},
\code{\link{glottosave}()}
}
\concept{<glottodata>}
\keyword{internal}
