% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget.R
\name{glottoget_path}
\alias{glottoget_path}
\title{Load user-provided glottodata}
\usage{
glottoget_path(
  filename = NULL,
  meta = FALSE,
  simple = TRUE,
  create = "glottodata"
)
}
\arguments{
\item{filename}{Path to glottodata file with extension (.xlsx .xls .gpkg .shp). If no filename is specified, an artificial dummy dataset will be created.}

\item{meta}{By default, meta sheets are not loaded. Use meta=TRUE if you want to include them.}

\item{create}{In case 'filename' is not specified, artificial dummy data will be created in glottodata format (specify create = "glottosubdata" to create data in glottosubdata format)}

\item{simplify}{By default, if only one sheet is loaded, the data will be returned as a data.frame (instead of placing the data inside a list of length 1)}
}
\value{

}
\description{
Load glottodadata/glottosubdata from a file, or create artificial dummy data.
}
\examples{
glottoget_path()
glottoget_path(filename = "glottodata.xlsx")
glottoget_path(filename = "glottodata.gpkg")
}
\seealso{
glottosave

Other <glottodata>: 
\code{\link{glottoget_remote}()},
\code{\link{glottoget}()},
\code{\link{glottosave}()}
}
\concept{<glottodata>}
