% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget_wals.R
\name{glottoget_wals}
\alias{glottoget_wals}
\title{Get WALS data}
\usage{
glottoget_wals(days = NULL, valuenames = NULL, paramnames = NULL)
}
\arguments{
\item{days}{After how many days should be checked for a new version?}

\item{valuenames}{Should names of the values be added instead of codes?}

\item{paramnames}{Should names of parameters (columns) be added instead of their codes?}
}
\description{
This function checks whether most recent version of WALS is locally available. If local version is outdated, the newest version will be downloaded.
}
\examples{
glottoget_wals()
}
\keyword{internal}
