% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottojoin.R
\name{glottojoin_base}
\alias{glottojoin_base}
\title{Join glottodata with glottobase}
\usage{
glottojoin_base(glottodata, id = NULL)
}
\arguments{
\item{glottodata}{User-provided glottodata}

\item{id}{Optional, name of column with glottocodes}
}
\description{
Keeps all rows from glottodata. Can be used to add features from glottobase.
}
\examples{
glottodata <- glottoget("demodata", meta = FALSE)
glottojoin_base(glottodata)
}
\keyword{internal}
