% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottojoin.R
\name{glottojoin_dist}
\alias{glottojoin_dist}
\title{Join glottodata with a dist object}
\usage{
glottojoin_dist(glottodata, id = NULL, dist, rm.na = FALSE)
}
\arguments{
\item{glottodata}{User-provided glottodata}

\item{id}{Column with IDs in \code{glottodata} that match names of dist object. By default, the "glottocode" column is used.}

\item{dist}{A dist object}

\item{rm.na}{Default is to keep NAs.}
}
\value{
Data frame
}
\description{
Join glottodata with a dist object
}
\examples{
distdf <- glottojoin_dist(glottodata = glottodata, dist = dist)

After joining, you can subset the distance columns by using the IDs:
distdf[, distdf$glottocode]
}
\keyword{internal}
