% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glottolog}
\alias{glottolog}
\title{Load glottolog data}
\format{
A data frame with XX rows and XX variables:
\describe{
  \item{var1}{explain}
  \item{var2}{explain}
  ...
}
}
\source{
\url{http://www.glottolog.org/
}
}
\usage{
glottolog
}
\description{
This is an internally stored version of glottolog data. It is included in the package as a fall-back option in case there is no internet connection.
}
\keyword{datasets}
