% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glottolog}
\alias{glottolog}
\title{This is an internally stored version of glottolog data. Use glottoget("glottolog", download = TRUE) to download the latest version}
\format{
A data frame based on glottolog
}
\source{
\url{https://glottolog.org/
}
}
\usage{
glottolog
}
\description{
This is an internally stored version of glottolog data. Use glottoget("glottolog", download = TRUE) to download the latest version
}
\keyword{internal}
\note{version: v4.7


Cite the source of the dataset as:


Hammarström, Harald & Forkel, Robert & Haspelmath, Martin & Bank, Sebastian. 2022. Glottolog 4.7. Leipzig: Max Planck Institute for Evolutionary Anthropology. (Available online at https://glottolog.org)


##------ Fri Jun 23 17:27:29 2023 ------##}
