% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glottolog}
\alias{glottolog}
\title{This is an internally stored version of glottolog data. It is included in the package as a fall-back option in case there is no internet connection.}
\format{
A data frame with XX rows and XX variables:
\describe{
  \item{id}{description}
  \item{name}{description}
  \item{macroarea}{description}
  \item{latitude}{description}
  \item{longitude}{description}
  \item{iso639p3code}{description}
  \item{countries}{description}
  \item{family_id}{description}
  \item{level}{description}
  \item{category}{description}
  \item{bookkeeping}{description}
  \item{classification}{description}
  \item{parent_id}{description}
  ...
}
}
\source{
\url{https://glottolog.org/
}
}
\usage{
glottolog
}
\description{
This is an internally stored version of glottolog data. It is included in the package as a fall-back option in case there is no internet connection.
}
\keyword{internal}
