% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glottolog}
\alias{glottolog}
\title{This is an internally stored version of glottolog data. Use glottoget("glottolog", download = TRUE) to download the latest version}
\format{
A data frame with XX rows and XX variables:
\describe{
  \item{id}{description}
  \item{name}{description}
  \item{macroarea}{description}
  \item{latitude}{description}
  \item{longitude}{description}
  \item{iso639p3code}{description}
  \item{countries}{description}
  \item{family_id}{description}
  \item{level}{description}
  \item{category}{description}
  \item{bookkeeping}{description}
  \item{classification}{description}
  \item{parent_id}{description}
  ...
}
}
\source{
\url{https://glottolog.org/
}
}
\usage{
glottolog
}
\description{
This is an internally stored version of glottolog data. Use glottoget("glottolog", download = TRUE) to download the latest version
}
\keyword{internal}
\note{version: v4.5


Cite the source of the dataset as:


Hammarström, Harald & Forkel, Robert & Haspelmath, Martin & Bank, Sebastian. 2021. Glottolog 4.5. Leipzig: Max Planck Institute for Evolutionary Anthropology. (Available online at https://glottolog.org)


##------ Tue Mar 15 11:44:11 2022 ------##}
