% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottologbooster.R
\name{glottologbooster}
\alias{glottologbooster}
\title{Glottologbooster: enhance glottolog data}
\usage{
glottologbooster(glottologdata, geoglot = TRUE)
}
\arguments{
\item{glottologdata}{}
}
\value{

}
\description{
Glottologbooster: enhance glottolog data
}
\examples{
glottologdata <- get_glottolog("glottolog")
glottologdata <- glottologbooster(glottologdata)
}
