% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottologbooster.R
\name{glottologbooster}
\alias{glottologbooster}
\title{Glottologbooster: enhance glottolog data}
\usage{
glottologbooster(glottologdata, geoglot = TRUE)
}
\arguments{
\item{glottologdata}{data from \href{https://glottolog.org/}{glottolog}, can be downloaded with \code{\link{glottoget("glottolog")}}}
}
\value{

}
\description{
Glottologbooster: enhance glottolog data
}
\examples{
glottologdata <- glottoget("glottolog")
glottologdata <- glottologbooster(glottologdata)
}
