% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottomap.R
\name{glottomap}
\alias{glottomap}
\title{glottomap}
\usage{
glottomap(
  glottodata,
  color = NULL,
  label = NULL,
  type = NULL,
  ptsize = NULL,
  transparency = NULL
)
}
\arguments{
\item{glottodata}{User-provided glottodata}

\item{color}{column name or index to be used to color features (optional)}

\item{label}{Column name or index to be used to label features (optional)}

\item{type}{One of: "static", "dynamic", or "interactive". Defaults to
"static" if nothing is provided.}

\item{ptsize}{Size of points between 0 and 1}

\item{transparency}{Transparency of points between 0 (very transparent) and 1 (not transparent)}
}
\value{

}
\description{
Create dynamic, static and interactive maps from glottodata
}
\examples{
glottopoints <- glottofilter(continent = "South America")
glottopols <- points2pols(glottopoints, method = "voronoi", continent = "South America")
glottomap(glottodata = glottopols, color = "family_size_rank")

glottodata <- get_glottobase()
families <- glottodata \%>\% dplyr::count(family_name, sort = TRUE)

# highlight 10 largest families:
glottodata <- glottospotlight(glottodata = glottodata, spotcol =
"family_name", spotlight = families$family_name[1:10], spotcontrast = "family_name", bgcontrast = "family_name")

# Or, place 10 largest families in background
glottodata <- glottospotlight(glottodata = glottodata, spotcol =
"family_name", spotlight = families$family_name[-c(1:10)], spotcontrast = "family_name", bgcontrast = "family_name")
glottomap(glottodata, color = "color")
}
