% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottomap.R
\name{glottomap}
\alias{glottomap}
\title{glottomap}
\usage{
glottomap(
  glottodata = NULL,
  color = NULL,
  label = NULL,
  type = NULL,
  ptsize = NULL,
  alpha = NULL,
  lbsize = NULL,
  palette = NULL,
  rivers = FALSE,
  nclass = NULL,
  numcat = FALSE,
  filename = NULL,
  ...
)
}
\arguments{
\item{glottodata}{Optional, user-provided glottodata. In case no glottodata is provided, you can pass arguments directly to glottofilter.}

\item{color}{glottovar, column name, or column index to be used to color features (optional). Run glottovars() to see glottovars}

\item{label}{glottovar, column name, or column index to be used to label features (optional). Run glottovars() to see glottovars}

\item{type}{One of: "static", "dynamic", or "interactive". Defaults to
"static" if nothing is provided.}

\item{ptsize}{Size of points between 0 and 1}

\item{alpha}{Transparency of points between 0 (very transparent) and 1 (not transparent)}

\item{lbsize}{Size of labels between 0 and 1}

\item{palette}{Color palette, see glottocolpal("all") for possible options, and run glottocolpal("turbo") to see what it looks like (replace it with palette name).
Alternatively, you could also run tmaptools::palette_explorer(), RColorBrewer::display.brewer.all(), ?viridisLite::viridis, or scales::show_col(viridisLite::viridis(n=20))}

\item{rivers}{Do you want to plot rivers (only for static maps)?}

\item{nclass}{Preferred number of classes (default is 5)}

\item{numcat}{Do numbers represent categories? For example, if your dataset consists of 0 and 1, you might want to set this to TRUE.}

\item{filename}{Optional filename if you want to save resulting map}

\item{...}{}
}
\value{

}
\description{
Create dynamic, static and interactive maps from glottodata.
}
\examples{
glottomap(country = "Netherlands")

glottopoints <- glottofilter(continent = "South America")
glottopols <- glottospace(glottopoints, method = "voronoi", continent = "South America")
glottomap(glottodata = glottopols, color = "family_size_rank")
glottomap(glottodata = glottopols, color = "family", palette = "turbo", type = "dynamic", label = "name")

glottodata <- glottoget()
families <- glottodata \%>\% dplyr::count(family, sort = TRUE)

# highlight 10 largest families:
glottodata <- glottospotlight(glottodata = glottodata, spotcol =
"family", spotlight = families$family[1:10], spotcontrast = "family", bgcontrast = "family")

# Or, place 10 largest families in background
glottodata <- glottospotlight(glottodata = glottodata, spotcol =
"family", spotlight = families$family[-c(1:10)], spotcontrast = "family", bgcontrast = "family")
glottomap(glottodata, color = "color")
}
\seealso{
geomap
}
\concept{<glottomap>}
