% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottomap.R
\name{glottomap}
\alias{glottomap}
\title{glottomap}
\usage{
glottomap(glottodata, colorby = NULL, label = NULL, type = NULL)
}
\arguments{
\item{glottodata}{User-provided glottodata (spatial)}

\item{colorby}{column name or index to be used to color features (optional)}

\item{label}{Column name or index to be used to label features (optional)}

\item{type}{One of: "static", "dynamic", "interactive", or "simple"}
}
\value{

}
\description{
Create dynamic, static and interactive maps from glottodata
}
\examples{
glottopoints <- glottofilter(continent = "South America")
glottopols <- points2pols(glottopoints, interpolation = "voronoi", continent = "South America")
glottomap_dynamic(glottodata = glottopols, label = "glottocode", colorby = "family_size_rank")
}
