% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottomap.R
\name{glottomap_dynamic}
\alias{glottomap_dynamic}
\title{Create a dynamic map with glottodata}
\usage{
glottomap_dynamic(
  glottodata,
  label = NULL,
  color = NULL,
  ptsize = NULL,
  alpha = NULL,
  palette = NULL,
  nclass = NULL,
  numcat = FALSE
)
}
\arguments{
\item{glottodata}{User-provided glottodata (either glottopoints or glottopols)}

\item{label}{Column name or index to be used to label features (optional)}

\item{color}{Column name or index to be used to color features (optional), or a color "black"}

\item{ptsize}{Size of points between 0 and 1}
}
\description{
Create a dynamic map with glottodata
}
\examples{
glottodata <- glottofilter(continent = "South America")
glottodata <- glottofilter(country = "Netherlands")
glottomap_dynamic(glottodata)
}
\keyword{internal}
