% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottomap.R
\name{glottomap_geodata}
\alias{glottomap_geodata}
\title{plotgeodata}
\usage{
glottomap_geodata(geodata)
}
\arguments{
\item{geodata}{}
}
\value{

}
\description{
plotgeodata
}
\examples{
nl <- get_geodata(download = "elevation", country = c("Netherlands"))
plotgeodata(nl)
}
\seealso{
Other <geodata>: 
\code{\link{extractgeodata}()},
\code{\link{extractraster}()},
\code{\link{get_geodata_download}()},
\code{\link{get_geodata_path}()},
\code{\link{get_geodata}()}
}
\concept{<geodata>}
\keyword{internal}
