% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottomap.R
\name{glottomap_rips_filt}
\alias{glottomap_rips_filt}
\title{Title}
\usage{
glottomap_rips_filt(glottodata, filt = NULL, dist_mtx = NULL, r = 0, maxscale)
}
\arguments{
\item{glottodata}{a glottodata with geometry type of `POINT`.}

\item{filt}{an object of Vietoris-Rips filtration, if it is `NULL`, it will be computed from glottodata.}

\item{dist_mtx}{a distance matrix corresponding to geographic distances of `glottodata`, if it is `NULL`, it will be computed from glottodata.}

\item{r}{the radius of buffers of all the points in glottodata, the unit of `r` is "100km".}

\item{maxscale}{a numeric number, maximum value of the rips filtration.}
}
\description{
Title
}
\examples{
\dontrun{
awk <- glottopoints[glottopoints$family == "Arawakan", ]
glottomap_rips_filt(glottodata = awk, r = 6, maxscale = 8)
}
}
