% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottomap.R
\name{glottomap_rips_filt}
\alias{glottomap_rips_filt}
\title{Title}
\usage{
glottomap_rips_filt(
  glottodata,
  r = 0,
  maxscale,
  is_animate = FALSE,
  length.out = 20,
  movie.name = "filtration.gif"
)
}
\arguments{
\item{glottodata}{a glottodata is an object of sf with geometry type as `POINT`}

\item{r}{a numerica number, the radius of buffers of all the points in glottodata, the default unit is "100km"}

\item{maxscale}{a numeric number, maximum value of the rips filtration, the default unit is "100km"}

\item{is_animate}{if TRUE, it will generate a GIF file, if FALSE, it will generate a tmap plot,
the default value is FALSE}

\item{length.out}{the amount of images to be generated in GIF file when `is_animate = TRUE`,
the default value is `20`}

\item{movie.name}{name of the GIF file, the default value is "filtration.gif"}
}
\value{
if `is_animate = FALSE` return a tmap, if `is_animate = TRUE` return a GIF file
}
\description{
Title
}
\examples{
glottopoints <- glottofilter(continent = "South America")
awk <- glottopoints[glottopoints$family == "Arawakan", ]
glottomap_rips_filt(glottodata = awk, r = 6, maxscale = 8)
\dontrun{
glottomap_rips_filt(glottodata = awk, r = 6, maxscale = 8, is_animate=TRUE)
}
}
