% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottomap.R
\name{glottomap_static}
\alias{glottomap_static}
\title{glottomap_static}
\usage{
glottomap_static(glottodata, label, colorby)
}
\arguments{
\item{glottodata}{User-provided glottodata}

\item{label}{Column name or index to be used to label features (optional)}

\item{colorby}{column name or index to be used to color features (optional)}
}
\value{

}
\description{
glottomap_static
}
\examples{
glottomap_dynamic(glottodata)
}
\keyword{internal}
