% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottomap.R
\name{glottomap_static}
\alias{glottomap_static}
\title{Create a static map with glottodata}
\usage{
glottomap_static(
  glottodata,
  projection = NULL,
  label = NULL,
  color = NULL,
  ptsize = 1,
  lbsize = NULL,
  alpha = 1,
  palette = NULL,
  rivers = FALSE,
  nclass = NULL,
  numcat = FALSE
)
}
\arguments{
\item{glottodata}{User-provided glottodata (either glottopoints or glottopols)}

\item{projection}{One of: 'eqarea' (equal-area Eckert IV, default), 'pacific' (Pacific-centered), or any other Coordinate Reference System, specified using an EPSG code (https://epsg.io/).}

\item{label}{Column name or index to be used to label features (optional)}

\item{color}{column name or index to be used to color features (optional), or a color "black"}

\item{ptsize}{Point size between 0 and 1}

\item{lbsize}{Label size between 0 an 1}

\item{alpha}{Transparency of points between 0 (very transparent) and 1 (not transparent)}

\item{palette}{Color palette, see glottocolpal("all") for possible options}

\item{rivers}{Do you want to plot rivers?}

\item{nclass}{Preferred number of classes (default is 5)}

\item{numcat}{Do numbers represent categories? For example, if your dataset consists of 0 and 1, you might want to set this to TRUE.}
}
\description{
This function returns a static map with glottodata. Data is projected using the equal-area Eckert IV projection (following McNew et al. 2018). See \url{https://en.wikipedia.org/wiki/Eckert_IV_projection}.
}
\examples{
glottodata <- glottofilter(continent = "South America")
glottodata <- glottofilter(country = c("Netherlands", "Germany", "Belgium") )
glottomap_static(glottodata)
}
\keyword{internal}
