% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoplot.R
\name{glottonmds}
\alias{glottonmds}
\title{Nonmetric Multidimensional Scaling}
\usage{
glottonmds(dist, k = 2, rm.na = FALSE)
}
\arguments{
\item{dist}{dist object or distance matrix}

\item{k}{Number of dimensions}

\item{rm.na}{Whether NAs should be removed (default is FALSE)}
}
\value{

}
\description{
Nonmetric Multidimensional Scaling
}
\examples{
glottodata <- glottoget_glottodata()
dist <- glottodist(glottodata = glottodata, structure = structure)
nmds <- glottonmds(dist = dist, k = 2)
}
\seealso{
Other <glottoplot>: 
\code{\link{glottonmds_scores}()}
}
\concept{<glottoplot>}
\keyword{internal}
