% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottonmds.R
\name{glottonmds}
\alias{glottonmds}
\title{Nonmetric Multidimensional Scaling for a glottodist object}
\usage{
glottonmds(glottodist = NULL, k = NULL, rm.na = FALSE, row2id = NULL)
}
\arguments{
\item{glottodist}{A glottodist object}

\item{k}{Number of dimensions. Either 2 or 3 for nmds.}

\item{rm.na}{Whether na's should be removed (default is FALSE)}

\item{row2id}{In case of nmds, specify what each row contains (either 'glottocode' or 'glottosubcode')}
}
\value{
a glottonmds object
}
\description{
Nonmetric Multidimensional Scaling for a glottodist object
}
