% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoplot.R
\name{glottoplot_nmds}
\alias{glottoplot_nmds}
\title{Plot nmds}
\usage{
glottoplot_nmds(
  nmds,
  scoresdata,
  colorby = NULL,
  sizeby = NULL,
  labelby = NULL,
  filename = NULL
)
}
\arguments{
\item{nmds}{An nmds object}

\item{scoresdata}{scoresdata}

\item{colorby}{colorby}

\item{sizeby}{sizeby}

\item{labelby}{label}

\item{filename}{optional filename if output should be saved.}
}
\value{

}
\description{
Plot nmds
}
\examples{
Join nmds scores with glottodata:
nmds <- glottonmds(glottodata, k = 2)
scores <- glottonmds_scores(nmds)
scoresdata <- join_glottobase(scores)

glottoplot_nmds(nmds = nmds, scoresdata = scoresdata, colorby = "family_name", sizeby = "isolate")
glottoplot_nmds(nmds = nmds, scoresdata = scoresdata, colorby = "isolate")
}
