% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoclean.R
\name{glottorecode_logical}
\alias{glottorecode_logical}
\title{Recode character columns to TRUE/FALSE}
\usage{
glottorecode_logical(glottodata, structure, totrue = NULL, tofalse = NULL)
}
\arguments{
\item{glottodata}{glottodata list}

\item{structure}{structure table}

\item{totrue}{values to recode to TRUE}

\item{tofalse}{values to recode to FALSE}
}
\description{
Recode character columns to TRUE/FALSE
}
\examples{
glottodata <- glottoget("demodata", meta = TRUE)
glottorecode_logical(glottodata, totrue = c("y", "Y", 1), tofalse = c("n", "N", 0),
structure = glottodata[["structure"]])

glottosubdata <- glottoget("demosubdata", meta = TRUE)
glottorecode_logical(glottosubdata, totrue = c("y", "Y", 1), tofalse = c("n", "N", 0),
structure = glottosubdata[["structure"]])
}
