% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottosave.R
\name{glottosave}
\alias{glottosave}
\title{Save glottodata in relevant format}
\usage{
glottosave(glottodata, filename = NULL)
}
\arguments{
\item{glottodata}{User-provided glottodata}

\item{filename}{Filename either with or without file extension}
}
\value{

}
\description{
If no file extention is provided, a sensible default is chosen. Dynamic maps
(tmap) are saved in .html format, static maps (tmap) are saved as .png.
Spatial data (sf) are saved as geopackage (.GPKG) by default, but .shp is
also possible.
}
\examples{

glottodata <- get_glottodata(meta = FALSE)
Saves as .xlsx
glottosave(glottodata, filename = "glottodata")

glottodata <- glottodata_makespatial(glottodata)
Saves as .GPKG
glottosave(glottodata, filename = "glottodata")

glottomap <- glottomap(glottodata)
Saves as .png
glottosave(glottomap, filename = "glottomap")

Saves as .html
glottomap <- glottomap(glottodata, type = "dynamic")
}
\seealso{
get_glottodata

Other <glottodata>: 
\code{\link{get_glottodata}()}
}
\concept{<glottodata>}
