% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottosave.R
\name{glottosave}
\alias{glottosave}
\title{Save any geoglot/glot object in relevant format}
\usage{
glottosave(object = NULL, filename = NULL)
}
\arguments{
\item{filename}{}
}
\value{

}
\description{
If no file extention is provided, a sensible default is chosen. Dynamic maps
(tmap) are saved in .html format, static maps (tmap) are saved as .png.
Spatial data (sf) are saved as geopackage (.GPKG) by default, but .shp is
also possible.
}
