% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottospace.R
\name{glottospace}
\alias{glottospace}
\title{Make glottodata spatial}
\usage{
glottospace(
  glottodata,
  method = NULL,
  radius = NULL,
  country = NULL,
  continent = NULL
)
}
\arguments{
\item{glottodata}{A glottodata table, or list of a glottodata table and metadata table(s)}

\item{method}{In case output should be polygons, interpolation method, either "buffer" or "voronoi" (synonymous with "thiessen")}

\item{radius}{In case interpolation method "buffer", the radius in km.}

\item{country}{Optionally mask output by country boundaries}

\item{continent}{Optionally mask output by continent boundaries}
}
\value{

}
\description{
This function takes glottodata (either with or without metadata) and turns it into spatial points or polygons.
}
\examples{
glottodata <- glottoget()
glottospacedata <- glottospace(glottodata)

}
