% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottospace.R
\name{glottospace_buffer}
\alias{glottospace_buffer}
\title{Create buffer around glottopoints}
\usage{
glottospace_buffer(glottodata, radius)
}
\arguments{
\item{glottodata}{spatial glottodata without metadata}

\item{radius}{radius in kilometers}
}
\value{

}
\description{
Create buffer around glottopoints
}
\examples{
glottodata <- glottofilter(country = "Netherlands")
}
\keyword{internal}
