% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottospace.R
\name{glottospace_thiessen}
\alias{glottospace_thiessen}
\title{Create Thiessen polygons}
\usage{
glottospace_thiessen(glottodata)
}
\arguments{
\item{glottodata}{spatial glottodata without metadata}
}
\value{

}
\description{
Create Thiessen polygons
}
\examples{
glottodata <- glottofilter(country = "Netherlands")
}
\keyword{internal}
