% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottostat.R
\name{glottostat_dist_permanova_mci}
\alias{glottostat_dist_permanova_mci}
\title{A temporary version of glottostat_dist_permanova}
\usage{
glottostat_dist_permanova_mci(
  glottodist = NULL,
  glottodata = NULL,
  comparison = NULL,
  sample = NULL,
  permutations = NULL,
  by = NULL
)
}
\arguments{
\item{glottodist}{a dist object}

\item{glottodata}{a glottodata}

\item{comparison}{comparision}

\item{sample}{sample}

\item{permutations}{permutations}

\item{by}{by}
}
\description{
A temporary version of glottostat_dist_permanova
}
