% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottosearch.R
\name{glottosubcode_valid}
\alias{glottosubcode_valid}
\title{Check whether glottosubcodes are valid.}
\usage{
glottosubcode_valid(glottosubcodes)
}
\arguments{
\item{glottosubcodes}{Character vector of glottosubcodes}
}
\value{
Gives warning in case there are issues, and invisibly returns TRUE otherwise.
}
\description{
This function checks whether a vector of glottosubcodes adheres to the
following form: glottocode_group_record. For example: abcd1234_aaa_0001,
abcd1234_aaa_0002, abcd1234_bbb_0001, abcd1234_bbb_0002.
}
\details{
Specifically, the function checks whether all glottocodes (which are part of the glotosubcodes) are valid, whether 'group' is a character, and whether 'record' is a number.
}
\examples{
glottosubcode_valid(c("yucu1253_aaa_0002", "abcd1234_aaa_0001"))
}
\seealso{
Other <glottocheck><glottosearch>: 
\code{\link{glottocode_exists}()},
\code{\link{glottocode_online}()},
\code{\link{glottosearch_exist}()}
}
\concept{<glottocheck><glottosearch>}
