% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grambank}
\alias{grambank}
\title{This is an internally stored version of grambank data (obtained from ...).}
\format{
An object of class \code{data.frame} with 2467 rows and 196 columns.
}
\usage{
grambank
}
\description{
This is an internally stored version of grambank data (obtained from ...).
}
\keyword{datasets}
