% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottojoin.R
\name{join_glottobase}
\alias{join_glottobase}
\title{Join glottodata with glottobase}
\usage{
join_glottobase(glottodata, id = NULL)
}
\arguments{
\item{glottodata}{User-provided glottodata}

\item{id}{Optional, name of column with glottocodes}
}
\value{

}
\description{
Keeps all rows from glottodata. Can be used to add features from glottobase.
}
\examples{
glottodata <- get_glottodata(meta = FALSE)
join_glottobase(glottodata)
}
