% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottojoin.R
\name{join_glottodata}
\alias{join_glottodata}
\title{Join two glottodata tables.}
\usage{
join_glottodata(glottodata, with, id = NULL)
}
\arguments{
\item{glottodata}{User-provided glottodata}

\item{with}{User-provided glottodata}

\item{id}{Optional, name of column with glottocodes}
}
\value{

}
\description{
Join two glottodata tables.
}
\examples{
glottodatax <- glottoget(meta = FALSE)
glottodatay <- glottoget(meta = FALSE)
join_glottodata(glottodatax, glottodatay)
}
\seealso{
glottodata_makespatial glottodata_addcoords
}
\keyword{internal}
