% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottojoin.R
\name{join_glottospace}
\alias{join_glottospace}
\title{Add coordinates to glottodata.}
\usage{
join_glottospace(glottodata, id = NULL)
}
\arguments{
\item{glottodata}{User-provided glottodata}

\item{id}{Optional, name of column with glottocodes}
}
\value{

}
\description{
Join glottodata with glottospace (keeps all rows from glottodata). It is recommended to use the function glottospace().
}
\examples{
glottodata <- glottoget_path(meta = FALSE)
join_glottospace(glottodata)
}
\seealso{
glottodata_makespatial glottodata_addcoords
}
\keyword{internal}
