% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodist.R
\name{meandist}
\alias{meandist}
\title{Calculate mean distance from one feature to all other features}
\usage{
meandist(dist)
}
\arguments{
\item{dist}{\code{dist} object}
}
\value{

}
\description{
Calculate mean distance from one feature to all other features
}
\examples{
meandist(dist)
}
\seealso{
Other <geodist><glottodist>: 
\code{\link{countwithinradius}()},
\code{\link{nearestdistance}()},
\code{\link{nearestfeature}()}
}
\concept{<geodist><glottodist>}
