% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotools.R
\name{mergevec}
\alias{mergevec}
\title{Merge multiple shapefiles}
\usage{
mergevec(paths, selection = NULL, outfile = NULL)
}
\arguments{
\item{paths}{Character vector of paths to shapefile}

\item{selection}{Optionally select data from merged shapefiles}

\item{outfile}{Optionally save file locally}

\item{overwrite}{Whether outfile should be overwritten if it exists}
}
\value{

}
\description{
Merge multiple shapefiles
}
\examples{
\dontrun{mergevec(paths = c("vectora.shp", "vectorb.shp"))}
}
\seealso{
Other <geodata><geotools>: 
\code{\link{extractgeodata}()},
\code{\link{extractraster}()},
\code{\link{rasterstack}()}
}
\concept{<geodata><geotools>}
\keyword{internal}
