% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotools.R
\name{mosaicras}
\alias{mosaicras}
\title{Mosaic raster tiles}
\usage{
mosaicras(
  path,
  outfile = "mosaic.tif",
  overwrite = FALSE,
  na.rm = TRUE,
  subdirs = TRUE
)
}
\arguments{
\item{path}{path to folder containing multiple raster files}

\item{outfile}{file path where to save resulting raster mosaic. Output will be save as .tif}

\item{overwrite}{Default is FALSE}

\item{na.rm}{Default is TRUE}

\item{subdirs}{By default, raster files in subdirectories are also included in the mosaic.}
}
\value{

}
\description{
Mosaic raster tiles
}
\examples{
\dontrun{mosaicras(path = "path/localpath", outfile = "mosaicv2.tif")}
}
\concept{<geotools><geodata>}
\keyword{internal}
