% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodist.R
\name{pointdist_bird}
\alias{pointdist_bird}
\title{Distances between spatial points, as the bird flies.}
\usage{
pointdist_bird(points, label)
}
\arguments{
\item{points}{Spatial point object}

\item{label}{Column name or number indicating how objects should be labeled}
}
\value{
dist object with distances (km) between all points
}
\description{
Calculate Euclidean distances between spatial points, as the bird flies (great circle distances/Haversine distances).
}
\examples{
pointdist_bird(points = points, label = "glottocode")
}
