% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wals}
\alias{wals}
\title{WALS data}
\format{
A data frame with XX rows and XX variables:
\describe{
  \item{glottocode}{description}
  \item{81A}{description}
  \item{82A}{description}
  \item{83A}{description}
  \item{87A}{description}
  \item{88A}{description}
  \item{89A}{description}
  \item{92A}{description}
  \item{93A}{description}
  \item{97A}{description}
  \item{112A}{description}
  \item{116A}{description}
  \item{143A}{description}
  \item{143E}{description}
  \item{143F}{description}
  \item{143G}{description}
  \item{144A}{description}
  \item{144B}{description}
  \item{144D}{description}
  \item{144H}{description}
  \item{144I}{description}
  \item{144J}{description}
  \item{144K}{description}
  \item{26A}{description}
  \item{33A}{description}
  \item{37A}{description}
  \item{38A}{description}
  \item{51A}{description}
  \item{57A}{description}
  \item{69A}{description}
  \item{85A}{description}
  \item{86A}{description}
  \item{90A}{description}
  \item{90C}{description}
  \item{95A}{description}
  \item{96A}{description}
  \item{101A}{description}
  \item{144L}{description}
  \item{144P}{description}
  \item{144Q}{description}
  \item{144R}{description}
  \item{144S}{description}
  \item{70A}{description}
  \item{71A}{description}
  \item{132A}{description}
  \item{133A}{description}
  \item{134A}{description}
  \item{135A}{description}
  \item{14A}{description}
  \item{15A}{description}
  \item{16A}{description}
  \item{17A}{description}
  \item{142A}{description}
  \item{1A}{description}
  \item{2A}{description}
  \item{3A}{description}
  \item{4A}{description}
  \item{5A}{description}
  \item{6A}{description}
  \item{7A}{description}
  \item{8A}{description}
  \item{9A}{description}
  \item{10A}{description}
  \item{11A}{description}
  \item{12A}{description}
  \item{18A}{description}
  \item{19A}{description}
  \item{20A}{description}
  \item{21A}{description}
  \item{21B}{description}
  \item{28A}{description}
  \item{29A}{description}
  \item{35A}{description}
  \item{39A}{description}
  \item{40A}{description}
  \item{44A}{description}
  \item{48A}{description}
  \item{49A}{description}
  \item{50A}{description}
  \item{65A}{description}
  \item{66A}{description}
  \item{67A}{description}
  \item{68A}{description}
  \item{77A}{description}
  \item{78A}{description}
  \item{84A}{description}
  \item{91A}{description}
  \item{94A}{description}
  \item{98A}{description}
  \item{99A}{description}
  \item{100A}{description}
  \item{102A}{description}
  \item{103A}{description}
  \item{104A}{description}
  \item{107A}{description}
  \item{111A}{description}
  \item{113A}{description}
  \item{114A}{description}
  \item{121A}{description}
  \item{129A}{description}
  \item{130A}{description}
  \item{13A}{description}
  \item{22A}{description}
  \item{23A}{description}
  \item{24A}{description}
  \item{25A}{description}
  \item{25B}{description}
  \item{27A}{description}
  \item{30A}{description}
  \item{31A}{description}
  \item{32A}{description}
  \item{34A}{description}
  \item{36A}{description}
  \item{41A}{description}
  \item{42A}{description}
  \item{46A}{description}
  \item{47A}{description}
  \item{52A}{description}
  \item{53A}{description}
  \item{54A}{description}
  \item{55A}{description}
  \item{58A}{description}
  \item{58B}{description}
  \item{59A}{description}
  \item{60A}{description}
  \item{61A}{description}
  \item{62A}{description}
  \item{63A}{description}
  \item{64A}{description}
  \item{72A}{description}
  \item{73A}{description}
  \item{74A}{description}
  \item{75A}{description}
  \item{76A}{description}
  \item{79A}{description}
  \item{79B}{description}
  \item{80A}{description}
  \item{90B}{description}
  \item{105A}{description}
  \item{106A}{description}
  \item{108A}{description}
  \item{108B}{description}
  \item{109A}{description}
  \item{109B}{description}
  \item{110A}{description}
  \item{115A}{description}
  \item{117A}{description}
  \item{118A}{description}
  \item{119A}{description}
  \item{120A}{description}
  \item{122A}{description}
  \item{123A}{description}
  \item{124A}{description}
  \item{125A}{description}
  \item{126A}{description}
  \item{127A}{description}
  \item{128A}{description}
  \item{131A}{description}
  \item{136A}{description}
  \item{136B}{description}
  \item{137A}{description}
  \item{137B}{description}
  \item{144M}{description}
  \item{138A}{description}
  \item{39B}{description}
  \item{45A}{description}
  \item{56A}{description}
  \item{143B}{description}
  \item{144F}{description}
  \item{144E}{description}
  \item{139A}{description}
  \item{140A}{description}
  \item{143C}{description}
  \item{144O}{description}
  \item{144G}{description}
  \item{43A}{description}
  \item{144C}{description}
  \item{144T}{description}
  \item{144V}{description}
  \item{144W}{description}
  \item{144X}{description}
  \item{81B}{description}
  \item{10B}{description}
  \item{130B}{description}
  \item{144N}{description}
  \item{144U}{description}
  \item{90D}{description}
  \item{144Y}{description}
  \item{90E}{description}
  \item{141A}{description}
  \item{90F}{description}
  \item{143D}{description}
  \item{90G}{description}
  \item{name}{description}
  \item{macroarea}{description}
  \item{isocode}{description}
  \item{countries}{description}
  \item{family_id}{description}
  \item{classification}{description}
  \item{parent_id}{description}
  \item{family}{description}
  \item{isolate}{description}
  \item{family_size}{description}
  \item{family_size_rank}{description}
  \item{country}{description}
  \item{continent}{description}
  \item{sovereignty}{description}
  \item{geometry}{description}
  ...
}
}
\source{
\url{https://wals.info/
}
}
\usage{
wals
}
\description{
This is an internally stored version of WALS data. It is included in the package as a fall-back option in case there is no internet connection.
}
\keyword{internal}
