% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget_wals.R
\name{wals_download_cldf}
\alias{wals_download_cldf}
\title{Download WALS data from zenodo, and select relevant data from cldf data}
\usage{
wals_download_cldf(valuenames = NULL, paramnames = NULL)
}
\arguments{
\item{valuenames}{Should names of the values be added instead of codes?}

\item{paramnames}{Should names of parameters (columns) be added instead of their codes?}
}
\description{
Download WALS data from zenodo, and select relevant data from cldf data
}
\keyword{internal}
