% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget_wals.R
\name{wals_loadlocal}
\alias{wals_loadlocal}
\title{Load locally stored WALS data and join with glottolog}
\usage{
wals_loadlocal()
}
\value{

}
\description{
Load locally stored WALS data and join with glottolog
}
\keyword{internal}
