% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottoget_wals.R
\name{wals_loadlocal_raw}
\alias{wals_loadlocal_raw}
\title{Load locally stored WALS data (without joining with glottolog)}
\usage{
wals_loadlocal_raw(valuenames = NULL, paramnames = NULL)
}
\arguments{
\item{valuenames}{Should names of the values be added instead of codes?}

\item{paramnames}{Should names of parameters (columns) be added instead of their codes?}
}
\value{

}
\description{
Load locally stored WALS data (without joining with glottolog)
}
\keyword{internal}
